% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattests.R
\name{mw_stattest_ageratio}
\alias{mw_stattest_ageratio}
\alias{mw_stattest_ageratio2}
\title{Test for statistical difference between the proportion of children aged 24 to
59 months old over those aged 6 to 23 months old}
\usage{
mw_stattest_ageratio(age, .expectedP = 0.66)

mw_stattest_ageratio2(age_cat, .expectedP = 0.66)
}
\arguments{
\item{age}{A \code{numeric} vector of child's age in months.}

\item{.expectedP}{The expected proportion of children aged 24 to 59 months
old over those aged 6 to 23 months old. By default, this is expected to be
0.66.}

\item{age_cat}{A \code{character} vector of child's age in categories. Code values
should be "6-23" and "24-59".}
}
\value{
A \code{list} object with three elements: \code{p} for p-value of the
difference between the observed and the expected proportion of children aged
24 to 59 months old over those aged 6 to 23 months old, \code{observedR} for the
observed ratio, and \code{observedP} for the observed proportion.
}
\description{
Calculate the observed age ratio of children aged 24 to 59 months old over
those aged 6 to 23 months old and test if there is a statistically
significant difference between the observed and the expected.
}
\details{
This function should be used specifically when assessing the quality of MUAC
data. For age ratio test of children aged 6 to 29 months old over 30 to 59
months old, as performed in the SMART plausibility check, use
\code{\link[nipnTK:ageRatioTest]{nipnTK::ageRatioTest()}} instead.
}
\examples{

mw_stattest_ageratio(
  age = anthro.02$age,
  .expectedP = 0.66
)


age <- ifelse(anthro.02$age < 24, "6-23", "24-59")

mw_stattest_ageratio2(
  age = age,
  .expectedP = 0.66
)

}
\references{
SMART Initiative. \emph{Updated MUAC data collection tool}. Available at:
\url{https://smartmethodology.org/survey-planning-tools/updated-muac-tool/}
}
