% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro.04}
\alias{anthro.04}
\title{A sample data from a community-based sentinel site with location anonymised}
\format{
A tibble of 2,192 × 6.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{analysis_unit} \tab Location wherein the assessment was conducted \cr
   \emph{cluster} \tab Primary sampling unit \cr
   \emph{sex} \tab Sex; "1" = boys, "2" = girls \cr
   \emph{age} \tab Calculated age in months \cr
   \emph{muac} \tab Mid-upper arm circumference in millimetres \cr
   \emph{oedema} \tab oedema; "n" = no oedema, "y" = yes oedema \cr
}
}
\source{
Anonymous
}
\usage{
anthro.04
}
\description{
Data herein was derived from population-based assessments in three locations
(analysis unit). Each unit presents distinct scenarios that requires
also-distinct handling during the prevalence analysis:
\itemize{
\item \emph{Unit A} has flawless data.
\item \emph{Unit B} has age-ratio-test results rated as problematic, with an observed
proportion of children aged 24-59 months < 0.66.
\item \emph{UnitC} has age-ratio-test results rated as problematic, with an observed
proportion of children aged 24-59 months >= 0.66.
}

This sample data is useful to demonstrate designed behaviours of
MUAC-prevalence functions to dealing with certain-and-expected flaws in the
data.
}
\examples{
anthro.04


}
\keyword{datasets}
