% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_muac.R
\name{apply_cdc_age_weighting}
\alias{apply_cdc_age_weighting}
\title{Apply the CDC/SMART prevalence weighting approach on MUAC data}
\usage{
apply_cdc_age_weighting(muac, age, .edema = NULL, status = c("sam", "mam"))
}
\arguments{
\item{muac}{A vector of class \code{integer} of MUAC values (in mm).}

\item{age}{A vector of class \code{double} of child's age in months.}

\item{.edema}{A vector of class \code{character} of edema. Code should be
"y" for presence and "n" for absence of bilateral edema. Default is \code{NULL}.}

\item{status}{A choice of the form of wasting to be defined.}
}
\value{
A vector of class \code{numeric} of length and size 1.
}
\description{
Calculate a weighted prevalence estimate of MUAC by adding the proportion of
children under 2 years to twice the proportion of children over 2 and then
dividing by 3.
}
\details{
This function is informed by the output of \code{\link[=age_ratio_test]{age_ratio_test()}}.
}
