% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_auditors.R
\name{check_plausibility_mfaz}
\alias{check_plausibility_mfaz}
\alias{check_plausibility_wfhz}
\alias{check_plausibility_muac}
\title{Audit the plausibility of WFHZ, MFAZ data and absolute MUAC values}
\usage{
check_plausibility_mfaz(df, sex, muac, age, flags, area)

check_plausibility_wfhz(df, sex, age, weight, height, flags, area)

check_plausibility_muac(df, flags, sex, muac)
}
\arguments{
\item{df}{A data frame yielded from \code{\link[=process_muac_data]{process_muac_data()}} for
\code{check_plausibility_mfaz()} and \code{check_plausibility_muac()}, and yielded from
\code{\link[=process_wfhz_data]{process_wfhz_data()}} for \code{check_plausibility_wfhz()}.}

\item{sex}{A vector holding codes on child's sex: 1 for boy and 2 for girl.}

\item{muac}{A numeric vector holding MUAC measurements (in centimeters).}

\item{age}{A numeric vector holding age in months.}

\item{flags}{A character vector holding on values on flagged observations.}

\item{area}{A character vector holding values on where was the data collected
and for which you want the analysis to be performed. If analysing data of just
one area, you will still have to supply the corresponding column to \code{area} in
\code{check_plausibility_mfaz()} or \code{check_plausibility_wfhz()}.}

\item{weight}{A numeric vector holding weight measurements (in kilograms).}

\item{height}{A numeric vector holding height measurements (in centimeters).}
}
\value{
A summarized table with the raw statistics and respective classification.
}
\description{
\code{check_plausibility_wfhz()}, \code{check_plausibility_mfaz()}, and
\code{check_plausibility_muac()} examines the plausibility of data through a
structured set of tests around sampling and measurement-related errors.
}
\examples{

## Audit the plausibility of MFAZ data ----

anthro.01 |>
process_age(
svdate = "dos",
birdate = "dob",
age = age
) |>
process_muac_data(
sex = sex,
age = "age",
muac = muac,
.recode_sex = TRUE,
.recode_muac = TRUE,
unit = "cm"
) |>
check_plausibility_mfaz(
flags = flag_mfaz,
sex = sex,
muac = muac,
age = age,
area = area
)

## Audit the plausibility of WFHZ ----

anthro.01 |>
process_age(
svdate = "dos",
birdate = "dob",
age = age
) |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
) |>
check_plausibility_wfhz(
sex = sex,
age = age,
weight = weight,
height = height,
flags = flag_wfhz,
area = area
)

## Audit the plausibility of the absolute MUAC values ----

anthro.01 |>
process_muac_data(
sex = sex,
muac = muac,
age = NULL,
.recode_sex = TRUE,
.recode_muac = FALSE,
unit = "none"
) |>
check_plausibility_muac(
flags = flag_muac,
sex = sex,
muac = muac
)

}
