% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_age.R
\name{calculate_age_in_months}
\alias{calculate_age_in_months}
\title{Calculate child's age in months}
\usage{
calculate_age_in_months(dos, dob)
}
\arguments{
\item{dos}{A vector of class \code{Date} for data collection date.}

\item{dob}{A vector of class \code{Date} for child's date of birth.}
}
\value{
A vector of class \code{double} for child's age in months with two decimal places.
Any value less than 6.0 and greater than or equal to 60.0 months will be set to \code{NA}.
}
\description{
Calculate child's age in months based on date of birth and the data collection date.
}
\keyword{internal}
