% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_definitions.R
\name{define_wasting_cases_muac}
\alias{define_wasting_cases_muac}
\alias{define_wasting_cases_whz}
\alias{define_wasting_cases_combined}
\alias{define_wasting}
\title{Define if an observation is wasted on the basis of the criteria
of WFHZ, absolute MUAC values and combined case-definition}
\usage{
define_wasting_cases_muac(muac, edema = NULL, cases = c("gam", "sam", "mam"))

define_wasting_cases_whz(zscore, edema = NULL, cases = c("gam", "sam", "mam"))

define_wasting_cases_combined(
  zscore,
  muac,
  edema = NULL,
  cases = c("cgam", "csam", "cmam")
)

define_wasting(
  df,
  zscore = NULL,
  muac = NULL,
  edema = NULL,
  base = c("wfhz", "muac", "combined")
)
}
\arguments{
\item{muac}{A numeric vector holding absolute MUAC values (in mm).}

\item{edema}{A character vector indicating if an observation has bilateral
edema or not. The codes are "y" for presence and "n" for absence of bilateral
edema. Default is \code{NULL}.}

\item{cases}{A choice of the form of wasting to be defined.}

\item{zscore}{A numeric vector holding WFHZ values (with 3 decimal places).}

\item{df}{A data frame containing the required variables.}

\item{base}{A choice of the criterion which the case-definition should be based
on.}
}
\value{
A numeric vector of the same length as the input vector, with dummy
values: 1 for yes wasted and 0 for not wasted. The meaning of the codes
changes depending on the form of wasting chosen. That is, if set \code{cases} to
\code{"sam"} the codes 1 would mean yes for severe wasting.
}
\description{
Define if an observation is wasted on the basis of the criteria
of WFHZ, absolute MUAC values and combined case-definition
}
\details{
Use \code{define_wasting()} to add the case-definitions into data frame.
}
\examples{
# MUAC-based case-definition ----
x <- anthro.02 |>
define_wasting(
muac = muac,
edema = edema,
base = "muac"
)
head(x)

# Weight-for-height based case-definition ----
x <- anthro.02 |>
define_wasting(
zscore = wfhz,
edema = edema,
base = "wfhz"
)
head(x)

# Combined case-definition ----
x <- anthro.02 |>
define_wasting(
zscore = wfhz,
muac = muac,
edema = edema,
base = "combined"
)
head(x)

}
