% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{check_sample_size}
\alias{check_sample_size}
\title{Check whether the IPC Acute Malnutrition sample size requirements were met}
\usage{
check_sample_size(df, .group, .data_type = c("survey", "screening", "ssite"))
}
\arguments{
\item{df}{A dataset of class \code{data.frame} to check.}

\item{.group}{A vector of class \code{integer} of the cluster ID's for survey,
screening or site ID's for screenings and sentinel sites.}

\item{.data_type}{A choice between "survey" for survey data, "screening" for
screening data or "ssite" for community-based sentinel site data.}
}
\value{
A summarised table of three columns: \code{groups} for the total number
of unique cluster or screening or site IDs; \code{n_obs} for the respective total
number of children; and \code{meet_ipc} for whether the IPC AMN requirements were met.
}
\description{
Verify whether the minimum sample size requirements for the area of analysis
were met, in accordance with the IPC Acute Malnutrition (IPC AMN) protocols.
}
\details{
\href{https://www.ipcinfo.org/ipcinfo-website/resources/ipc-manual/en/}{The IPC Manual}.
}
\examples{

anthro.01 |>
dplyr::group_by(area) |>
check_sample_size(
.group = cluster,
.data_type = "survey"
)

}
