% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_raters.R
\name{classify_overall_quality}
\alias{classify_overall_quality}
\title{Rate the overall data quality}
\usage{
classify_overall_quality(df)
}
\arguments{
\item{df}{A data frame containing a vector with the quality scores yielded
from \code{\link[=compute_quality_score]{compute_quality_score()}}.}
}
\value{
A character vector of the same length with a new column called
\code{quality_class}.
}
\description{
\code{classify_overall_quality()} informs you about the overall quality of the data
by rating the overall quality score in "Excellent", "Good", "Acceptable" and
"Problematic".
}
\examples{
## A sample data ----

df <- data.frame(
quality_score = 29
)

## Apply the function ----
classify_overall_quality(df)

}
