% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_raters.R
\name{classify_skew_kurt}
\alias{classify_skew_kurt}
\title{Rate the acceptability of the skewness and kurtosis test results}
\usage{
classify_skew_kurt(sk)
}
\arguments{
\item{sk}{A vector of class \code{double} for skewness or kurtosis test results.}
}
\value{
A vector of class \code{character} of the same length as \code{sk} for the
acceptability rate.
}
\description{
Rate the acceptability of the skewness and kurtosis test results
}
