% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_definitions.R
\name{classify_wasting_for_cdc_approach}
\alias{classify_wasting_for_cdc_approach}
\title{Classify wasting into severe or moderate wasting to be used in the
SMART MUAC tool weighting approach}
\usage{
classify_wasting_for_cdc_approach(muac, .edema = NULL)
}
\arguments{
\item{muac}{A numeric vector holding absolute MUAC values (in mm).}

\item{.edema}{Optional. A character vector indicating if an observation has
bilateral edema or not. The codes are "y" for presence and "n" for absence of
bilateral edema.}
}
\value{
A character vector of the same length as the input indicating if a
child is severe or moderate wasted or not wasted.
}
\description{
Classify wasting into severe or moderate wasting to be used in the
SMART MUAC tool weighting approach
}
