% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_combined.R
\name{compute_pps_based_combined_prevalence}
\alias{compute_pps_based_combined_prevalence}
\alias{compute_combined_prevalence}
\title{Compute combined prevalence of acute malnutrition}
\usage{
compute_pps_based_combined_prevalence(
  df,
  .wt = NULL,
  .edema = NULL,
  .summary_by
)

compute_combined_prevalence(df, .wt = NULL, .edema = NULL, .summary_by = NULL)
}
\arguments{
\item{df}{A data frame object returned by \code{\link[=process_muac_data]{process_muac_data()}} and \code{\link[=process_wfhz_data]{process_wfhz_data()}}.
Both wranglers need to be used to prepare data to be used
\code{compute_combined_prevalence()}. The order of which comes first does not matter,
however, since the muac data processor transforms MUAC values into centimeters, those
need to be put back into millimeter. This can be achieved my using \code{\link[=recode_muac]{recode_muac()}} inside
\code{\link[dplyr:mutate]{dplyr::mutate()}} or \code{\link[base:transform]{base::transform()}}.}

\item{.wt}{A numeric vector holding final survey weights. When set to \code{NULL},
the function assumes self weighted survey, as in the ENA for SMART software;
Otherwise when supplied, weighted analysis is computed.}

\item{.edema}{A character vector indicating if an observation has bilateral
edema or not. The codes are "y" for presence and "n" for absence of bilateral
edema. Default is \code{NULL}.}

\item{.summary_by}{A character vector containing data on the geographical areas
where the data was collected and for which the analysis should be performed at.}
}
\value{
A table with the descriptive statistics about wasting.
}
\description{
\code{compute_combined_prevalence()} is a handy function for calculating the
combined prevalence of wasting also in with the complex sample design
properties inherent to surveys.
}
\details{
The concept of "combined flags" is introduced in this function. It consists of
taking the \code{flag_wfhz} and \code{flag_mfaz} vectors, generated from the MUAC and
WFHZ wranglers, and checking if any value in either vector is flagged. If flagged,
the value is marked as a flag in the "cflags" vector; otherwise, it is not flagged
(see table below). This ensures that all flagged observations from both WFHZ
and MFAZ data are excluded from the combined prevalence analysis.\tabular{ccc}{
   \strong{flag_wfhz} \tab \strong{flag_mfaz} \tab \strong{cflags} \cr
   1 \tab 0 \tab 1 \cr
   0 \tab 1 \tab 1 \cr
   0 \tab 0 \tab 0 \cr
}
}
\examples{

## When .summary_by and .wt are set to NULL ----
p <- compute_combined_prevalence(
df = anthro.02,
.wt = NULL,
.edema = edema,
.summary_by = NULL
)

print(p)

## When .wt is not set to NULL ----
x <- compute_combined_prevalence(
df = anthro.02,
.wt = "wtfactor",
.edema = edema,
.summary_by = NULL
)

print(x)

## When working on data frame with multiple survey areas ----
s <- anthro.03 |>
process_age(age = age) |>
process_muac_data(
sex = sex,
muac = muac,
age = "age",
.recode_sex = TRUE,
.recode_muac = TRUE,
unit = "cm"
) |>
dplyr::mutate(muac = recode_muac(muac, unit = "mm")) |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE) |>
compute_combined_prevalence(
.edema = edema,
.summary_by = district
)

print(s)

}
