% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_muac.R
\name{compute_muac_prevalence}
\alias{compute_muac_prevalence}
\title{This is to be revised}
\usage{
compute_muac_prevalence(df, .wt = NULL, .edema = NULL, .summary_by = NULL)
}
\description{
This is to be revised
}
\examples{

## An example of application of `compute_muac_prevalence()` ----

### When .summary.by = NULL ----

x <- compute_muac_prevalence(
df = anthro.04,
.wt = NULL,
.edema = edema,
.summary_by = NULL
)

print(x)

### When .summary_by is not set to NULL ----

p <- compute_muac_prevalence(
df = anthro.04,
.wt = NULL,
.edema = edema,
.summary_by = province
)

print(p)

}
