% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scorers.R
\name{compute_quality_score}
\alias{compute_quality_score}
\title{Get the overall quality score for WFHZ and MFAZ}
\usage{
compute_quality_score(df, type = c("mfaz", "whz"))
}
\arguments{
\item{df}{A data frame containing individual test quality scores.}

\item{type}{The method you wish to get the overall quality score for.
A choice between "mfaz" and "wfhz".}
}
\value{
A vector named \code{"quality_score"} with the overall quality score.
}
\description{
\code{compute_quality_score()} calculates the overall score of the quality of the
data for both WFHZ and MFAZ.
}
\examples{

## A sample data ----

df <- data.frame(
flagged_class = "Excellent",
age_ratio_class = "Good",
sex_ratio_class = "Problematic",
dps_class = "Excellent",
sd_class = "Excellent",
skew_class = "Good",
kurt_class = "Acceptable"
)

## Apply the function ----
compute_quality_score(df, type = "mfaz")

}
