% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_muac.R
\name{compute_weighted_prevalence}
\alias{compute_weighted_prevalence}
\title{Apply the CDC/SMART prevalence weighting approach on MUAC data}
\usage{
compute_weighted_prevalence(df, .edema = NULL, .summary_by = NULL)
}
\arguments{
\item{df}{An already wrangled dataset object of class \code{data.frame} to use.}

\item{.edema}{A vector of class \code{character} of edema. Code should be
"y" for presence and "n" for absence of bilateral edema. Default is \code{NULL}.}

\item{.summary_by}{A vector of class \code{character} of the geographical areas
where the data was collected and for which the analysis should be performed.}
}
\value{
A table of class \code{data.frame} of dimensions that vary based on
\code{.summary_by}, containing the results.
}
\description{
Apply the CDC/SMART prevalence weighting approach on MUAC data
}
