% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_age.R
\name{convert_month_to_days}
\alias{convert_month_to_days}
\title{Calculate child's age in days}
\usage{
convert_month_to_days(x)
}
\arguments{
\item{x}{A double vector of child's age in months.}
}
\value{
A double vector of the same length as \code{x} of age in days.
}
\description{
Calculate child's age in days
}
\keyword{internal}
