% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipccheckr-package.R
\docType{package}
\name{ipccheckr-package}
\alias{ipccheckr}
\alias{ipccheckr-package}
\title{ipccheckr: Utilities for analysing children's nutritional status}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A streamlined and comprehensive implementation of SMART Methodology guidelines for data quality checks and prevalence analyses, with enhanced efficiency for handling large datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nutriverse/mwana}
  \item Report bugs at \url{https://github.com/nutriverse/mwana/issues}
}

}
\author{
\strong{Maintainer}: Tomás Zaba \email{tomas.zaba@outlook.com} (\href{https://orcid.org/0000-0002-7079-3574}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ernest Guevarra (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [copyright holder]
}

}
\keyword{internal}
