% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_helpers.R
\name{mod_call_plausibility_checkers}
\alias{mod_call_plausibility_checkers}
\title{Call mwana's plausibility checkers into the module server depending on
user specifications in the UI}
\usage{
mod_call_plausibility_checkers(
  df,
  age = NULL,
  sex,
  muac = NULL,
  weight = NULL,
  height = NULL,
  flags,
  area1,
  area2,
  area3,
  .for = c("wfhz", "muac", "mfaz")
)
}
\description{
Call mwana's plausibility checkers into the module server depending on
user specifications in the UI
}
\keyword{internal}
