% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_helpers.R
\name{mod_call_prevalence_function_screening2}
\alias{mod_call_prevalence_function_screening2}
\title{Call mwana's prevalence functions into the module server depending on
user specifications in the UI}
\usage{
mod_call_prevalence_function_screening2(
  df,
  age_cat,
  muac,
  oedema = NULL,
  area1,
  area2,
  area3
)
}
\description{
Call mwana's prevalence functions into the module server depending on
user specifications in the UI
}
\keyword{internal}
