% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_helpers.R
\name{mod_display_input_variables_ipccheck}
\alias{mod_display_input_variables_ipccheck}
\title{Display input variables dynamically, according to UI for screening}
\usage{
mod_display_input_variables_ipccheck(vars, ns)
}
\description{
Display input variables dynamically, according to UI for screening
}
\keyword{internal}
