% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_helpers.R
\name{mod_display_input_variables_survey}
\alias{mod_display_input_variables_survey}
\title{Displays input variables dynamically, according to UI for screening}
\usage{
mod_display_input_variables_survey(vars, ns)
}
\description{
Displays input variables dynamically, according to UI for screening
}
\keyword{internal}
