% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_muac.R
\name{mw_estimate_prev_wasting_muac}
\alias{mw_estimate_prev_wasting_muac}
\title{This is to be revised}
\usage{
mw_estimate_prev_wasting_muac(df, wt = NULL, edema = NULL, .by = NULL)
}
\description{
This is to be revised
}
\examples{

## An example of application of `mw_estimate_prev_wasting_muac()` ----

### When .summary.by = NULL ----

prev <- mw_estimate_prev_wasting_muac(
  df = anthro.04,
  wt = NULL,
  edema = edema,
  .by = NULL
)

print(prev)

### When .by is not set to NULL ----

p <- mw_estimate_prev_wasting_muac(
  df = anthro.04,
  wt = NULL,
  edema = edema,
  .by = province
)

print(p)

}
