% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev-wasting-muac.R
\name{mw_estimate_prevalence_muac}
\alias{mw_estimate_prevalence_muac}
\title{Estimate the prevalence of wasting based on MUAC for survey data}
\usage{
mw_estimate_prevalence_muac(df, age, muac, wt = NULL, oedema = NULL, ...)
}
\arguments{
\item{df}{A \code{tibble} object produced by \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}} and
\code{\link[=mw_wrangle_age]{mw_wrangle_age()}} functions. Note that MUAC values in \code{df}
must be in millimetres after using \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}}. Also, \code{df}
must have a variable called \code{cluster} wherein the primary sampling unit
identifiers are stored.}

\item{age}{A vector of class \code{double} of child's age in months.}

\item{muac}{A \code{numeric} or \code{integer} vector of raw MUAC values. The
measurement unit should be millimetres.}

\item{wt}{A vector of class \code{double} of the survey sampling weights. Default
is NULL, which assumes a self-weighted survey, the case of SMART surveys.
Otherwise, a weighted analysis is implemented.}

\item{oedema}{A \code{character} vector for presence of nutritional oedema Code
values should be "y" for presence and "n" for absence. Default is NULL.}

\item{...}{A vector of class \code{character}, specifying the categories for which
the analysis should be summarised for. Usually geographical areas. More than
one vector can be specified.}
}
\value{
A summary \code{tibble} for the descriptive statistics about wasting based
on MUAC, with confidence intervals.
}
\description{
Estimate the prevalence of wasting based on MUAC and/or nutritional oedema.
The function allows users to estimate prevalence in accordance with complex
sample design properties, such as accounting for survey sample weights when
needed or applicable.

It first evaluates the quality of the data to determine the appropriate
prevalence-analysis flow to be employed. Quality is evaluated by estimating
the observed proportion of children aged 24-59 months of the total children in
the dataset, then it estimates the p-value for the difference between the
above-mentioned category against the expected (0.66) and rates it.

If age ratio test is "problematic" and the proportion of children aged 24-59
months is < 0.66, age-weighting approach is applied to prevalence estimation,
to account for the over-representation of younger children in the sample;
otherwise, a non-age-weighted prevalence is estimated.
}
\details{
A typical user analysis workflow is expected to begin with data quality checks,
followed by a thorough review, and only thereafter proceed to prevalence
estimation. This sequence places the user in the strongest position to assess
whether the resulting prevalence estimates are reliable.

Outliers are identified using SMART flagging criteria applied to MFAZ, and are
excluded from the prevalence estimation.
}
\examples{

## Ungrouped analysis ----
anthro.04 |>
  mw_wrangle_age(age = age) |>
  mw_wrangle_muac(
    muac = muac,
    .recode_muac = TRUE,
    .to = "cm",
    age = age,
    sex = sex,
    .recode_sex = FALSE
  ) |>
  transform(muac = recode_muac(muac, "mm")) |>
  mw_estimate_prevalence_muac(
    muac = muac,
    age = age,
    wt = NULL,
    oedema = oedema,
    analysis_unit
  )

}
\references{
SMART Initiative (no date). \emph{Updated MUAC data collection tool}. Available at:
\url{https://smartmethodology.org/survey-planning-tools/updated-muac-tool/}
}
\seealso{
\code{\link[=mw_estimate_age_weighted_prev_muac]{mw_estimate_age_weighted_prev_muac()}} \code{\link[=mw_estimate_prevalence_mfaz]{mw_estimate_prevalence_mfaz()}}
\code{\link[=mw_estimate_prevalence_screening]{mw_estimate_prevalence_screening()}}
}
