% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_wfhz.R
\name{mw_estimate_prevalence_wfhz}
\alias{mw_estimate_prevalence_wfhz}
\title{Estimate the prevalence of wasting based on weight-for-height z-scores (WFHZ)}
\usage{
mw_estimate_prevalence_wfhz(df, wt = NULL, edema = NULL, ...)
}
\arguments{
\item{df}{A \code{tibble} object that has been produced by the \code{\link[=mw_wrangle_wfhz]{mw_wrangle_wfhz()}}
functions. The \code{df} should have a variable named \code{cluster} for the primary
sampling unit identifiers.}

\item{wt}{A vector of class \code{double} of the survey sampling weights. Default
is NULL which assumes a self-weighted survey as is the case for a survey
sample selected proportional to population size (i.e., SMART survey sample).
Otherwise, a weighted analysis is implemented.}

\item{edema}{A \code{character} vector for presence of nutritional edema coded as
"y" for presence of nutritional edema and "n" for absence of nutritional
edema. Default is NULL.}

\item{...}{A vector of class \code{character}, specifying the categories for which
the analysis should be summarised for. Usually geographical areas. More than
one vector can be specified.}
}
\value{
A summary \code{tibble} for the descriptive statistics about wasting.
}
\description{
Calculate the prevalence estimates of wasting based on z-scores of
weight-for-height and/or nutritional edema. The function allows users to
estimate prevalence in accordance with complex sample design properties such
as accounting for survey sample weights when needed or applicable. The
quality of the data is first evaluated by calculating and rating the standard
deviation of WFHZ. Standard approach to prevalence estimation is calculated
only when the standard deviation of MFAZ is rated as not problematic. If
the standard deviation is problematic, prevalence is estimated using the
PROBIT estimator. Outliers are detected based on SMART flagging criteria.
Identified outliers are then excluded before prevalence estimation is
performed.
}
\examples{
## When .by = NULL ----
### Start off by wrangling the data ----
data <- mw_wrangle_wfhz(
  df = anthro.03,
  sex = sex,
  weight = weight,
  height = height,
  .recode_sex = TRUE
)

### Now run the prevalence function ----
mw_estimate_prevalence_wfhz(
  df = data,
  wt = NULL,
  edema = edema
)

## Now when .by is not set to NULL ----
mw_estimate_prevalence_wfhz(
  df = data,
  wt = NULL,
  edema = edema,
  district
)

## When a weighted analysis is needed ----
mw_estimate_prevalence_wfhz(
  df = anthro.02,
  wt = wtfactor,
  edema = edema,
  province
)

}
