% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_check_mfaz.R
\name{mw_neat_output_mfaz}
\alias{mw_neat_output_mfaz}
\title{Clean and format the output table returned from the MFAZ plausibility check
for improved clarity and readability}
\usage{
mw_neat_output_mfaz(df)
}
\arguments{
\item{df}{A data frame containing the summary table returned by this package's
MFAZ plausibility check function. Must be of class \code{data.frame}.}
}
\value{
A data frame of the same length and width as \code{df}, with column names and
values formatted for clarity.
}
\description{
Clean and format the output table returned from the MFAZ plausibility check
for improved clarity and readability. It converts scientific notations to standard
notations, round values and rename columns to meaningful names.
}
\examples{
## First wrangle age data ----
data <- mw_wrangle_age(
df = anthro.01,
dos = dos,
dob = dob,
age = age,
.decimals = 2
)

## Then wrangle MUAC data ----
data_mfaz <- mw_wrangle_muac(
df = data,
sex = sex,
age = age,
muac = muac,
.recode_sex = TRUE,
.recode_muac = TRUE,
.to = "cm"
)

## Then run plausibility check ----
pl <- mw_plausibility_check_mfaz(
df = data_mfaz,
flags = flag_mfaz,
sex = sex,
muac = muac,
age = age
)

## Now neat the output table ----
mw_neat_output_mfaz(df = pl)

}
