% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_check_muac.R
\name{mw_neat_output_muac}
\alias{mw_neat_output_muac}
\title{Clean and format the output table returned from the MUAC plausibility check
for improved clarity and readability.}
\usage{
mw_neat_output_muac(df)
}
\arguments{
\item{df}{An object of class \code{data.frame} returned by this package's
plausibility checker for raw MUAC data, containing the summarized results to be
formatted.}
}
\value{
A \code{data.frame} object of the same length and width as \code{df}, with column names and
values formatted for clarity and readability.
}
\description{
Clean and format the output table returned from the plausibility check of raw
MUAC data for improved clarity and readability. It converts scientific notations
to standard notations, round values and rename columns to meaningful names.
}
\examples{
## First wranlge MUAC data ----
df_muac <- mw_wrangle_muac(
  df = anthro.01,
  sex = sex,
  muac = muac,
  age = NULL,
  .recode_sex = TRUE,
  .recode_muac = FALSE,
  .to = "none"
)

## Then run the plausibility check ----
pl_muac <- mw_plausibility_check_muac(
  df = df_muac,
  flags = flag_muac,
  sex = sex,
  muac = muac
)

## Neat the output table ----

mw_neat_output_muac(df = pl_muac)

}
