% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_check_wfhz.R
\name{mw_neat_output_wfhz}
\alias{mw_neat_output_wfhz}
\title{Clean and format the output table returned from the WFHZ plausibility check
for improved clarity and readability}
\usage{
mw_neat_output_wfhz(df)
}
\arguments{
\item{df}{An object of class \code{data.frame} returned by this package's
plausibility checker for WFHZ data, containing the summarized results to be
formatted.}
}
\value{
A \code{data.frame} object of the same length and width as \code{df}, with column names and
values formatted for clarity and readability.
}
\description{
Clean and format the output table returned from the WFHZ plausibility check
for improved clarity and readability. It converts scientific notations to standard
notations, round values and rename columns to meaningful names.
}
\examples{
## First wrangle age data ----
data <- mw_wrangle_age(
  df = anthro.01,
  dos = dos,
  dob = dob,
  age = age,
  .decimals = 2
)

## Then wrangle WFHZ data ----
data_wfhz <- mw_wrangle_wfhz(
  df = data,
  sex = sex,
  weight = weight,
  height = height,
  .recode_sex = TRUE
)

## Now run the plausibility check ----
pl <- mw_plausibility_check_wfhz(
  df = data_wfhz,
  sex = sex,
  age = age,
  weight = weight,
  height = height,
  flags = flag_wfhz
)

## Now neat the output table ----
mw_neat_output_wfhz(df = pl)


}
