% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility-check-mfaz.R
\name{mw_plausibility_check_mfaz}
\alias{mw_plausibility_check_mfaz}
\title{Check the plausibility and acceptability of MUAC-for-age z-score (MFAZ) data}
\usage{
mw_plausibility_check_mfaz(df, sex, muac, age, flags, ...)
}
\arguments{
\item{df}{A \code{data.frame} object to check.}

\item{sex}{A \code{numeric} vector for child's sex with 1 = males and 2 = females.}

\item{muac}{A \code{numeric} vector of child's MUAC in centimetres.}

\item{age}{A vector of class \code{double} of child's age in months.}

\item{flags}{A \code{numeric} vector of flagged records.}

\item{...}{A vector of class \code{character}, specifying the categories for which
the analysis should be summarised for. Usually geographical areas. More than
one vector can be specified.}
}
\value{
A single-row summary \code{tibble} with columns containing the plausibility
check results. If ungrouped analysis, the output will consist of 17 columns
and one row; otherwise, the number of columns will vary according to the number
vectors specified, and the number of rows to the categories within the grouping
variables.
}
\description{
Check the overall plausibility and acceptability of MFAZ data through a
structured test suite encompassing checks for sampling and
measurement-related biases in the dataset. This test suite follows the
recommendation made by Bilukha & Kianian (2023) on the plausibility of
constructing a comprehensive plausibility check for MUAC data similar to
weight-for-height z-score to evaluate its acceptability when age values are
available in the dataset.

The function works on a \code{data.frame} returned from wrangling functions for
age and for MUAC-for-age z-score data available from this package.
}
\details{
Whilst the function uses the same checks and criteria as those for
weight-for-height z-scores in the SMART plausibility check, the percent of
flagged records is evaluated using different cut-off points, with a maximum
acceptability of 2.0\% as shown below:\tabular{cccc}{
   \strong{Excellent} \tab \strong{Good} \tab \strong{Acceptable} \tab \strong{Problematic} \cr
   0.0 - 1.0 \tab >1.0 - 1.5 \tab >1.5 - 2.0 \tab >2.0 \cr
}
}
\examples{
## First wrangle age data ----
data <- mw_wrangle_age(
  df = anthro.01,
  dos = dos,
  dob = dob,
  age = age,
  .decimals = 2
)

## Then wrangle MUAC data ----
data_muac <- mw_wrangle_muac(
  df = data,
  sex = sex,
  age = age,
  muac = muac,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  .to = "cm"
)

## And finally run plausibility check ----
mw_plausibility_check_mfaz(
  df = data_muac,
  flags = flag_mfaz,
  sex = sex,
  muac = muac,
  age = age,
  area, team
)

}
\references{
Bilukha, O., & Kianian, B. (2023). Considerations for assessment of measurement
quality of mid‐upper arm circumference data in anthropometric surveys and
mass nutritional screenings conducted in humanitarian and refugee settings.
\emph{Maternal & Child Nutrition}, 19, e13478. \url{https://onlinelibrary.wiley.com/doi/10.1111/mcn.13478}

SMART Initiative (2017). \emph{Standardized Monitoring and Assessment for Relief
and Transition}. Manual 2.0. Available at: \url{https://smartmethodology.org}.
}
\seealso{
\code{\link[=mw_wrangle_age]{mw_wrangle_age()}} \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}} \code{\link[=mw_stattest_ageratio]{mw_stattest_ageratio()}}
\code{\link[=flag_outliers]{flag_outliers()}}
}
