% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility-check-muac.R
\name{mw_plausibility_check_muac}
\alias{mw_plausibility_check_muac}
\title{Check the plausibility and acceptability of raw MUAC data}
\usage{
mw_plausibility_check_muac(df, sex, muac, flags, ...)
}
\arguments{
\item{df}{A \code{data.frame} object to check. It must have been wrangled using
the \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}} function.}

\item{sex}{A \code{numeric} vector for child's sex with 1 = males and 2 = females.}

\item{muac}{A vector of class \code{double} of child's MUAC in centimetres.}

\item{flags}{A \code{numeric} vector of flagged records.}

\item{...}{A vector of class \code{character}, specifying the categories for which
the analysis should be summarised for. Usually geographical areas. More than
one vector can be specified.}
}
\value{
A single-row summary \code{tibble} with columns containing the plausibility
check results. If ungrouped analysis, the output will consist of nine columns
and one row; otherwise, the number of columns will vary according to the number
vectors specified, and the number of rows to the categories within the grouping
variables.
}
\description{
Check the overall plausibility and acceptability of raw MUAC data
through a structured test suite encompassing checks for sampling and
measurement-related biases in the dataset. The test suite in this function
follows the recommendation made by Bilukha & Kianian (2023).
}
\details{
Cut-off points used for the percent of flagged records:\tabular{cccc}{
   \strong{Excellent} \tab \strong{Good} \tab \strong{Acceptable} \tab \strong{Problematic} \cr
   0.0 - 1.0 \tab >1.0 - 1.5 \tab >1.5 - 2.0 \tab >2.0 \cr
}
}
\examples{
## First wrangle MUAC data ----
df_muac <- mw_wrangle_muac(
  df = anthro.01,
  sex = sex,
  muac = muac,
  age = NULL,
  .recode_sex = TRUE,
  .recode_muac = FALSE,
  .to = "none"
)

## Then run the plausibility check ----
mw_plausibility_check_muac(
  df = df_muac,
  flags = flag_muac,
  sex = sex,
  muac = muac,
  area, team # group analysis by survey area and by survey team
)

}
\references{
Bilukha, O., & Kianian, B. (2023). Considerations for assessment of measurement
quality of mid‐upper arm circumference data in anthropometric surveys and
mass nutritional screenings conducted in humanitarian and refugee settings.
\emph{Maternal & Child Nutrition}, 19, e13478. \url{https://onlinelibrary.wiley.com/doi/10.1111/mcn.13478}

SMART Initiative (2017). \emph{Standardized Monitoring and Assessment for Relief
and Transition}. Manual 2.0. Available at: \url{https://smartmethodology.org}.
}
\seealso{
\code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}} \code{\link[=flag_outliers]{flag_outliers()}}
}
