% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_check_wfhz.R
\name{mw_plausibility_check_wfhz}
\alias{mw_plausibility_check_wfhz}
\title{Check the plausibility and acceptability of a weight-for-height data}
\usage{
mw_plausibility_check_wfhz(df, sex, age, weight, height, flags)
}
\arguments{
\item{df}{A dataset object of class \code{data.frame} to check. It should have been
wrangled using this package's wranglers.}

\item{sex}{A vector of class \code{numeric} of child's sex: 1 for boy and 2 for girl.}

\item{age}{A vector of class \code{double} of child's age in months.}

\item{weight}{A vector of class \code{double} of child's weight in kilograms.}

\item{height}{A vector of class \code{double} of child's height in centimeters.}

\item{flags}{A vector of class \code{numeric} of flagged observations.}
}
\description{
Verify the overall acceptability of the data through a set of
structured tests around sampling and measurement-related biases in the data.
}
