% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{mw_run_app}
\alias{mw_run_app}
\title{Initialise built-in Shiny application}
\usage{
mw_run_app(package = "mwana")
}
\arguments{
\item{package}{package name (\code{mwana}).}
}
\description{
Initialise built-in Shiny application
}
\examples{
if (interactive()) mw_run_app()

}
