% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_age.R
\name{mw_wrangle_age}
\alias{mw_wrangle_age}
\title{Wrangle child's age}
\usage{
mw_wrangle_age(df, dos = NULL, dob = NULL, age, .decimals = 2)
}
\arguments{
\item{df}{A \code{data.frame} object to wrangle age from.}

\item{dos}{A \code{Date} vector of dates when data collection was conducted.
Default is NULL.}

\item{dob}{A \code{Date} vector of dates of birth of child. Default is NULL.}

\item{age}{A \code{numeric} vector of child's age in months. In most cases this
will be estimated using local event calendars or calculated age in months
based on date of data collection and date of birth of child.}

\item{.decimals}{The number of decimal places to round off age to. Default is
2.}
}
\value{
A \code{tibble} based on \code{df}. The variable \code{age} will be automatically
filled in each row where age value was missing and both the child's
date of birth and the date of data collection are available. Rows where \code{age}
is less than 6.0 and greater than or equal to 60.0 months old will be set to
NA. Additionally, a new variable named \code{age_days} of class \code{double} for
calculated age of child in days is added to \code{df}.
}
\description{
Wrangle child's age for downstream analysis. This includes calculating age
in months based on the date of data collection and the child's date of birth,
and setting to NA the age values that are less than 6.0 and greater than or
equal to 60.0 months old.
}
\examples{

## A sample data ----
df <- data.frame(
  surv_date = as.Date(c(
    "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01"
  )),
  birth_date = as.Date(c(
    "2019-01-01", NA, "2018-03-20", "2019-11-05", "2021-04-25"
  )),
  age = c(NA, 36, NA, NA, NA)
)

## Apply the function ----
mw_wrangle_age(
  df = df,
  dos = surv_date,
  dob = birth_date,
  age = age,
  .decimals = 3
)

}
