% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle-muac.R
\name{mw_wrangle_muac}
\alias{mw_wrangle_muac}
\title{Wrangle MUAC data}
\usage{
mw_wrangle_muac(
  df,
  sex,
  muac,
  age = NULL,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  .to = c("cm", "mm", "none"),
  .decimals = 3
)
}
\arguments{
\item{df}{A \code{data.frame} object to wrangle data from.}

\item{sex}{A \code{numeric} or \code{character} vector of child's sex. Code values
should only be 1 or "m" for males and 2 or "f" for females.}

\item{muac}{A \code{numeric} vector of child's age in months.}

\item{age}{A \code{numeric} vector of child's age in months. Default is NULL.}

\item{.recode_sex}{Logical. Set to TRUE if the values for \code{sex} are not coded
as 1 (for males) or 2 (for females). Otherwise, set to FALSE (default).}

\item{.recode_muac}{Logical. Set to TRUE if the values for raw MUAC should be
converted to either centimetres or millimetres. Otherwise, set to FALSE
(default)}

\item{.to}{A choice of the measuring unit to convert MUAC values into. Can be
"cm" for centimetres, "mm" for millimetres, or "none" to leave as it is.}

\item{.decimals}{The number of decimal places to use for z-score outputs.
Default is 3.}
}
\value{
A \code{tibble} based on \code{df}. If \code{age = NULL}, \code{flag_muac} variable for
detected MUAC outliers based on raw MUAC is added to \code{df}. Otherwise,
variables named \code{mfaz} for child's MFAZ and \code{flag_mfaz} for detected outliers
based on SMART guidelines are added to \code{df}.
}
\description{
Calculate z-scores for MUAC-for-age (MFAZ) and identify outliers based on
the SMART methodology. When age is not supplied, only outliers are detected
from the raw MUAC values. The function only works after age has gone through
\code{\link[=mw_wrangle_age]{mw_wrangle_age()}}.
}
\examples{
## When age is available, wrangle it first before calling the function ----
w <- mw_wrangle_age(
  df = anthro.02,
  dos = NULL,
  dob = NULL,
  age = age,
  .decimals = 2
)

### Then apply the function to wrangle MUAC data ----
mw_wrangle_muac(
  df = w,
  sex = sex,
  age = age,
  muac = muac,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  .to = "cm",
  .decimals = 3
)

## When age is not available ----
mw_wrangle_muac(
  df = anthro.02,
  sex = sex,
  age = NULL,
  muac = muac,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  .to = "cm",
  .decimals = 3
)

}
\references{
Bilukha, O., & Kianian, B. (2023). Considerations for assessment of measurement
quality of mid‐upper arm circumference data in anthropometric surveys and
mass nutritional screenings conducted in humanitarian and refugee settings.
\emph{Maternal & Child Nutrition}, 19, e13478. \url{https://onlinelibrary.wiley.com/doi/10.1111/mcn.13478}

SMART Initiative (2017). \emph{Standardized Monitoring and Assessment for Relief
and Transition}. Manual 2.0. Available at: \url{https://smartmethodology.org}.
}
\seealso{
\code{\link[=flag_outliers]{flag_outliers()}} \code{\link[=remove_flags]{remove_flags()}} \code{\link[=mw_wrangle_age]{mw_wrangle_age()}}
}
