% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_wfhz.R
\name{mw_wrangle_wfhz}
\alias{mw_wrangle_wfhz}
\title{Wrangle weight-for-height data}
\usage{
mw_wrangle_wfhz(df, sex, weight, height, .recode_sex = TRUE, .decimals = 3)
}
\arguments{
\item{df}{A \code{data.frame} object to wrangle data from.}

\item{sex}{A \code{numeric} or \code{character} vector of child's sex. Code values
should only be 1 or "m" for males and 2 or "f" for females.}

\item{weight}{A vector of class \code{double} of child's weight in kilograms.}

\item{height}{A vector of class \code{double} of child's height in centimeters.}

\item{.recode_sex}{Logical. Set to TRUE if the values for \code{sex} are not coded
as 1 (for males) or 2 (for females). Otherwise, set to FALSE (default).}

\item{.decimals}{The number of decimal places to use for z-score outputs.
Default is 3.}
}
\value{
A data frame based on \code{df} with new variables named \code{wfhz} for
child's WFHZ and \code{flag_wfhz} for detected outliers added.
}
\description{
Calculate z-scores for weight-for-height (WFHZ) and identify outliers based
on the SMART methodology.
}
\examples{
mw_wrangle_wfhz(
  df = anthro.01,
  sex = sex,
  weight = weight,
  height = height,
  .recode_sex = TRUE,
  .decimals = 2
)

}
\references{
SMART Initiative (2017). \emph{Standardized Monitoring and Assessment for Relief}
\emph{and Transition}. Manual 2.0. Available at: \url{https://smartmethodology.org}.
}
\seealso{
\code{\link[=flag_outliers]{flag_outliers()}} \code{\link[=remove_flags]{remove_flags()}}
}
