% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwana-package.R
\docType{package}
\name{mwana-package}
\alias{mwana}
\alias{mwana-package}
\title{mwana: An Efficient Workflow for Plausibility Checks and Prevalence Analysis of Wasting in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A simple and streamlined workflow for plausibility checks and prevalence analysis of wasting based on the Standardized Monitoring and Assessment of Relief and Transition (SMART) Methodology \url{https://smartmethodology.org/}, with application in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mphimo/mwana}
  \item \url{https://mphimo.github.io/mwana/}
  \item Report bugs at \url{https://github.com/mphimo/mwana/issues}
}

}
\author{
\strong{Maintainer}: Tomás Zaba \email{tomas.zaba@outlook.com} (\href{https://orcid.org/0000-0002-7079-3574}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Tomás Zaba \email{tomas.zaba@outlook.com} (\href{https://orcid.org/0000-0002-7079-3574}{ORCID}) [copyright holder]
  \item Ernest Guevarra (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [copyright holder]
  \item Mark Myatt
}

}
\keyword{internal}
