% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flag_outliers}
\alias{flag_outliers}
\alias{remove_flags}
\title{Identify, flag, and remove outliers}
\usage{
flag_outliers(x, .from = c("zscores", "raw_muac"))

remove_flags(x, .from = c("zscores", "raw_muac"))
}
\arguments{
\item{x}{A \code{numeric} vector of WFHZ, MFAZ, HFAZ, WFAZ or raw MUAC values.
Raw MUAC values should be in millimetre units.}

\item{.from}{Either "zscores" or "raw_muac" for type of data to flag
outliers from.}
}
\value{
An vector of the same length as \code{x} of flagged records coded as
\code{1} for a flagged record and \code{0} for a non-flagged record.
}
\description{
Identify outlier z-scores for weight-for-height (WFHZ) and MUAC-for-age
(MFAZ) following the SMART methodology. The function can also be used to
detect outliers for height-for-age (HFAZ) and weight-for-age (WFAZ) z-scores
following the same approach.

For flagging z-scores, z-scores that deviate substantially from the sample's
z-score mean are considered outliers and are unlikely to reflect accurate
measurements. For raw MUAC, values that are less than 100 millimetres or
greater than 200 millimetres are considered outliers as recommended by
Bilukha & Kianian (2023). Including these values in the analysis could
compromise the accuracy of the resulting estimates.

To remove outliers, their values are set to NA rather than removing the
record from the dataset. This process is also called \emph{censoring}. By
assigning NA values to these outliers, they can be effectively removed
during statistical operations with functions that allow for removal of NA
values such as \code{\link[=mean]{mean()}} for getting the mean value or \code{\link[=sd]{sd()}} for getting the
standard deviation.
}
\examples{
## Sample data of raw MUAC values ----
x <- anthro.01$muac

## Apply the function with `.from` set to "raw_muac" ----
m <- flag_outliers(x, .from = "raw_muac")
head(m)

## Sample data of z-scores (be it WFHZ, MFAZ, HFAZ or WFAZ) ----
x <- anthro.02$mfaz

# Apply the function with `.from` set to "zscores" ----
z <- flag_outliers(x, .from = "zscores")
tail(z)

## With `.from` set to "zscores" ----
z <- remove_flags(
  x = wfhz.01$wfhz,
  .from = "zscores"
)

head(z)

## With `.from` set to "raw_muac" ----
m <- remove_flags(
  x = mfaz.01$muac,
  .from = "raw_muac"
)

tail(m)

}
\references{
Bilukha, O., & Kianian, B. (2023). Considerations for assessment of measurement
quality of mid‐upper arm circumference data in anthropometric surveys and
mass nutritional screenings conducted in humanitarian and refugee settings.
\emph{Maternal & Child Nutrition}, 19, e13478. Available at \url{https://onlinelibrary.wiley.com/doi/10.1111/mcn.13478}

SMART Initiative (2017). \emph{Standardized Monitoring and Assessment for Relief
and Transition}. Manual 2.0. Available at: \url{https://smartmethodology.org}.
}
