% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_auditors.R
\name{check_plausibility_mfaz}
\alias{check_plausibility_mfaz}
\alias{check_plausibility_wfhz}
\alias{check_plausibility_muac}
\title{Check the plausibility of the data}
\usage{
check_plausibility_mfaz(df, sex, muac, age, flags, area)

check_plausibility_wfhz(df, sex, age, weight, height, flags, area)

check_plausibility_muac(df, flags, sex, muac)
}
\arguments{
\item{df}{A dataset object of class \code{data.frame} to check. It should have been
wrangled using this package's wranglers.}

\item{sex}{A vector of class \code{numeric} of child's sex: 1 for boy and 2 for girl.}

\item{muac}{A vector of class \code{double} of child's MUAC in centimeters.}

\item{age}{A vector of class \code{double} of child's age in months.}

\item{flags}{A vector of class \code{numeric} of flagged observations.}

\item{area}{A vector of class \code{character} of the geographical location where
data was collected and to which the analysis should be aggregated at.}

\item{weight}{A vector of class \code{double} of child's weight in kilograms.}

\item{height}{A vector of class \code{double} of child's height in centimeters.}
}
\value{
A summarised \code{data.frame} of plausibility test results and their
respective acceptability ratings.
}
\description{
Verify the overall acceptability of the data through a set of
structured tests around sampling and measurement-related biases in the data.
}
\examples{

## Check the plausibility of WFHZ data ----

anthro.01 |>
process_age(
svdate = "dos",
birdate = "dob",
age = age
) |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
) |>
check_plausibility_wfhz(
sex = sex,
age = age,
weight = weight,
height = height,
flags = flag_wfhz,
area = area
)

## Check the plausibility of MFAZ data ----

anthro.01 |>
process_age(
svdate = "dos",
birdate = "dob",
age = age
) |>
process_muac_data(
sex = sex,
age = "age",
muac = muac,
.recode_sex = TRUE,
.recode_muac = TRUE,
unit = "cm"
) |>
check_plausibility_mfaz(
flags = flag_mfaz,
sex = sex,
muac = muac,
age = age,
area = area
)

## Check the plausibility of the absolute MUAC values ----

anthro.01 |>
process_muac_data(
sex = sex,
muac = muac,
age = NULL,
.recode_sex = TRUE,
.recode_muac = FALSE,
unit = "none"
) |>
check_plausibility_muac(
flags = flag_muac,
sex = sex,
muac = muac
)

}
