% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_muac.R
\name{mw_estimate_smart_age_wt}
\alias{mw_estimate_smart_age_wt}
\alias{mw_estimate_prevalence_muac}
\title{Estimate the prevalence of wasting based on MUAC}
\usage{
mw_estimate_smart_age_wt(df, edema = NULL, .by = NULL)

mw_estimate_prevalence_muac(df, wt = NULL, edema = NULL, .by = NULL)
}
\arguments{
\item{df}{A data set object of class \code{data.frame} to use. This must have been
wrangled using this package's wrangling function for MUAC data.
The function uses a variable name called \code{cluster} where the primary sampling unit IDs
are stored. Make sure the data set has this variable and its name has been
renamed to \code{cluster}, otherwise the function will error and terminate the execution.}

\item{edema}{A vector of class \code{character} of edema. Code should be
"y" for presence and "n" for absence of bilateral edema. Default is \code{NULL}.}

\item{.by}{A vector of class \code{character} or \code{numeric} of the geographical areas
or respective IDs for where the data was collected and for which the analysis
should be summarized at.}

\item{wt}{A vector of class \code{double} of the final survey weights. Default is
\code{NULL} assuming a self weighted survey, as in the ENA for SMART software;
otherwise, when a vector of weights if supplied, weighted analysis is done.}
}
\value{
A summarized table of class \code{data.frame} of the descriptive
statistics about wasting.
}
\description{
Calculate the prevalence estimates of wasting based on MUAC and/or bilateral
edema. The function first learns learns about the acceptability of the age
ratio test and of the standard deviation of muac-for-age z-scores (MFAZ) in
the data set, and then, depending on the observed results, it picks an analysis
workflow that best fits the data, between standard cluster-based survey design
including and weighted analysis, application of the age-weighting approach
as in the SMART MUAC Tool.

The function has considerable usefulness when working with a multiple-area
data set. Definition of workflow fit is done row-wise.
}
\examples{
## An application of `mw_estimate_smart_age_wt()` ----
.data <- anthro.04 |>
  subset(province == "Province 2")

mw_estimate_smart_age_wt(
  df = .data,
  edema = edema,
  .by = NULL
)

## When .by = NULL ----
mw_estimate_prevalence_muac(
  df = anthro.04,
  wt = NULL,
  edema = edema,
  .by = NULL
)

## When .by is not set to NULL ----
p <- mw_estimate_prevalence_muac(
  df = anthro.04,
  wt = NULL,
  edema = edema,
  .by = province
)

}
\references{
SMART Initiative (no date). \emph{Updated MUAC data collection tool}. Available at:
\url{https://smartmethodology.org/survey-planning-tools/updated-muac-tool/}
}
\seealso{
\code{\link[=mw_estimate_smart_age_wt]{mw_estimate_smart_age_wt()}}
}
