% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_raters.R
\name{rate_agesex_ratio}
\alias{rate_agesex_ratio}
\title{Rate the acceptability of the age and sex ratio test p-values}
\usage{
rate_agesex_ratio(p)
}
\arguments{
\item{p}{A vector of class \code{double} of the age or sex ratio test p-values.}
}
\value{
A \code{character} vector with the same length as \code{p} for the
acceptability rate.
}
\description{
Rate the acceptability of the age and sex ratio test p-values
}
\keyword{internal}
