% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_raters.R
\name{rate_propof_flagged}
\alias{rate_propof_flagged}
\title{Rate the acceptability of the proportion of flagged records}
\usage{
rate_propof_flagged(p, .in = c("mfaz", "wfhz", "raw_muac"))
}
\arguments{
\item{p}{A vector of class \code{double} of the proportions of flagged records in
the data set.}

\item{.in}{Specifies the data set where the rating should be done. Can be
"wfhz", "mfaz", or "raw_muac". Default to "wfhz".}
}
\value{
A vector of class \code{factor} with the same length as \code{p} for the
acceptability rate.
}
\description{
Rate the acceptability of the proportion of flagged records in WFHZ, MFAZ,
and raw MUAC data following the SMART methodology criteria.
}
\keyword{internal}
