% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_raters.R
\name{rate_skewkurt}
\alias{rate_skewkurt}
\title{Rate the acceptability of the skewness and kurtosis test results}
\usage{
rate_skewkurt(sk)
}
\arguments{
\item{sk}{A vector of class \code{double} for skewness or kurtosis test results.}
}
\value{
A vector of class \code{factor} with the same length as \code{sk} for the
acceptability rate.
}
\description{
Rate the acceptability of the skewness and kurtosis test results
}
\keyword{internal}
