% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scorers.R
\name{score_overall_quality}
\alias{score_overall_quality}
\title{Get the overall acceptability score from the acceptability classification scores}
\usage{
score_overall_quality(
  cl_flags,
  cl_sex,
  cl_age,
  cl_dps_m = NULL,
  cl_dps_w = NULL,
  cl_dps_h = NULL,
  cl_std,
  cl_skw,
  cl_kurt,
  .for = c("wfhz", "mfaz")
)
}
\arguments{
\item{.for}{A choice between "wfhz" and "mfaz" for the basis on which the
calculations should be made.}

\item{df}{A dataset object of class \code{data.frame} to calculate from.}
}
\value{
A \code{data.frame} based on \code{df} with a new column named \code{"quality_score"}
for the overall of acceptability (of quality) score.
}
\description{
Calculate the total amount of penalty points based on each plausibility test
result acceptability classification for WFHZ and MFAZ.
}
\keyword{internal}
