% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scorers.R
\name{score_std_flags}
\alias{score_std_flags}
\alias{score_agesexr_dps}
\alias{score_skewkurt}
\title{Score the acceptability rating of the check results that constitutes the
plausibility check suite}
\usage{
score_std_flags(x)

score_agesexr_dps(x)

score_skewkurt(x)
}
\arguments{
\item{x}{A \code{character} vector of the acceptability rate of a given check.
'}
}
\value{
An \code{integer} vector with the same length as \code{x} of the acceptability
score.
}
\description{
Attribute a score, also known as \emph{penalty point}, for a given rate of
acceptability of the standard deviation, proportion of flagged records,
age and sex ratio, skewness, kurtosis and digit preference score check
results. The scoring criteria and thresholds follows the standards in the SMART
plausibility check.
}
\references{
SMART Initiative (2017). \emph{Standardized Monitoring and Assessment for Relief
and Transition}. Manual 2.0. Available at: \url{https://smartmethodology.org}.
}
\keyword{internal}
