% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wfhz.01}
\alias{wfhz.01}
\title{A sample SMART survey data with WFHZ standard deviation rated as problematic}
\format{
A tibble with 303 rows and 6 columns.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{cluster} \tab Primary sampling unit \cr
   \emph{sex} \tab Sex, "m" = boys, "f" = girls \cr
   \emph{age} \tab calculated age in months with two decimal places \cr
   \emph{edema} \tab Edema, "n" = no, "y" = yes \cr
   \emph{wfhz} \tab MUAC-for-age z-scores with 3 decimal places \cr
   \emph{flag_wfhz} \tab Flagged observations. 1=flagged, 0=not flagged \cr
}
}
\source{
Anonymous
}
\usage{
wfhz.01
}
\description{
A sample SMART survey data with WFHZ standard deviation rated as problematic
}
\examples{
wfhz.01


}
\keyword{datasets}
