% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_muac.R, R/wrangle_wfhz.R
\name{mw_wrangle_muac}
\alias{mw_wrangle_muac}
\alias{mw_wrangle_wfhz}
\title{Wrangle weight-for-height and MUAC data}
\usage{
mw_wrangle_muac(
  df,
  sex,
  muac,
  age = NULL,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  .to = c("cm", "mm", "none"),
  .decimals = 3
)

mw_wrangle_wfhz(df, sex, weight, height, .recode_sex = TRUE, .decimals = 3)
}
\arguments{
\item{df}{A dataset of class \code{data.frame} to wrangle data from.}

\item{sex}{A \code{numeric} or \code{character} vector of child's sex. Code values should
be 1 or "m" for boy and 2 or "f" for girl. The variable name must be sex,
otherwise it will not work.}

\item{muac}{A vector of class \code{numeric} of child's age in months.}

\item{age}{A vector of class \code{numeric} of child's age in months.}

\item{.recode_sex}{Logical. Default is \code{FALSE}. Setting to \code{TRUE} assumes that
the sex variable is a character vector of values "m" for boys and "f" for girls
and will recode them to 1 and 2 respectively.}

\item{.recode_muac}{Logical. Default is \code{FALSE}. Set to \code{TRUE} if MUAC values
should be converted to either centimeters or millimeters.}

\item{.to}{A choice of the unit to which the MUAC values should be converted.
"cm" for centimeters, "mm" for millimeters and "none" to leave as it is.}

\item{.decimals}{The number of decimals places the z-scores should have.
Default is 3.}

\item{weight}{A vector of class \code{double} of child's weight in kilograms.}

\item{height}{A vector of class \code{double} of child's height in centimeters.}
}
\value{
A data frame based on \code{df}. New variables named \code{wfhz} and
\code{flag_wfhz}, of child's weight-for-height z-scores and flags, or \code{mfaz} and
\code{flag_mfaz}, of child's MUAC-for-age z-scores and flags, will be created. For
MUAC, when age is not supplied only \code{flag_muac} variable is created.
This refers to flags based on the absolute MUAC values as recommended by
\href{https://doi.org/10.1111/mcn.13478}{Bilukha, O., & Kianian, B. (2023).}.
}
\description{
This function performs data wrangling by calculating the weight-for-height
and the MUAC-for-age z-scores, followed by the detection and flagging of outliers.
For MUAC data, if age is not supplied, z-scores do not get computed. In such
cases, outlier detection and flagging will be based on the absolute MUAC values.
}
\details{
The flagging criterion used for the WFHZ and MFAZ is as in
\href{https://smartmethodology.org/}{SMART plausibility check}. A fixed flagging
criterion is used for the absolute MUAC values. This is as recommended by
\href{https://doi.org/10.1111/mcn.13478}{Bilukha, O., & Kianian, B. (2023).}
}
\examples{

## An example application of `mw_wrangle_wfhz()` ----
anthro.01 |>
  mw_wrangle_wfhz(
    sex = sex,
    weight = weight,
    height = height,
    .recode_sex = TRUE,
    .decimals = 2
  )

## An example application of `mw_wrangle_muac()` ----
### Sample data ----
df <- data.frame(
  survey_date = as.Date(c(
    "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01"
  )),
  birth_date = as.Date(c(
    "2019-01-01", NA, "2018-03-20", "2019-11-05", "2021-04-25"
  )),
  age = c(NA, 36, NA, NA, NA),
  sex = c("m", "f", "m", "m", "f"),
  muac = c(110, 130, 300, 123, 125)
)

### The application of the function ----
df |>
  mw_wrangle_age(
    dos = survey_date,
    dob = birth_date,
    age = age,
    .decimals = 2
  ) |>
  mw_wrangle_muac(
    sex = sex,
    age = "age",
    muac = muac,
    .recode_sex = TRUE,
    .recode_muac = TRUE,
    .to = "cm",
    .decimals = 3
  )

}
