% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro.03}
\alias{anthro.03}
\title{A sample data of district level SMART surveys conducted in Mozambique}
\format{
A tibble of 943 x 9.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{district} \tab The administrative unit (admin 1) where data was collected. \cr
   \emph{cluster} \tab Primary sampling unit \cr
   \emph{team} \tab Survey teams \cr
   \emph{sex} \tab Sex, "m" = boys, "f" = girls \cr
   \emph{age} \tab calculated age in months with two decimal places \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{edema} \tab Edema, "n" = no, "y" = yes \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
}
}
\usage{
anthro.03
}
\description{
\code{anthro.03} contains survey data of four districts. Each district's dataset
presents distinct data quality scenarios that requires tailored prevalence
analysis approach. Two districts show a problematic WFHZ standard deviation
whilst the remaining are all within range.

This sample data demonstrates the use of prevalence functions on multi-area
survey data, where variations in the standard deviation ratings exist.
As a result, different analytical approaches are required for each area
to ensure accurate estimation.
}
\examples{
anthro.03


}
\keyword{datasets}
