% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro.03}
\alias{anthro.03}
\title{A sample data of district level SMART surveys conducted in Mozambique}
\format{
A tibble of 943 x 9.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{district} \tab Survey location \cr
   \emph{cluster} \tab Primary sampling unit \cr
   \emph{team} \tab Survey teams \cr
   \emph{sex} \tab Sex; "m" = boys, "f" = girls \cr
   \emph{age} \tab Calculated age in months with two decimal places \cr
   \emph{weight} \tab Weight in kilograms \cr
   \emph{height} \tab Height in centimetres \cr
   \emph{edema} \tab Edema; "n" = no edema, "y" = with edema \cr
   \emph{muac} \tab Mid-upper arm circumference in millimetres \cr
}
}
\source{
Anonymous
}
\usage{
anthro.03
}
\description{
\code{anthro.03} contains survey data of four districts. Each district data set
presents distinct data quality scenarios that require a specific prevalence
analysis approach. Data from two districts have a problematic WFHZ standard
deviation. The data from the remaining two districts are all within range.

This sample data is useful to demonstrate the use of the prevalence functions
on a multiple-domain survey data where there can be variations in the rating
of acceptability of the standard deviation, hence requiring different
analytical approach for each survey domain to ensure accurate estimation.
}
\examples{
anthro.03


}
\keyword{datasets}
