% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_definitions.R
\name{define_wasting_cases_muac}
\alias{define_wasting_cases_muac}
\alias{define_wasting_cases_whz}
\alias{define_wasting_cases_combined}
\alias{define_wasting}
\title{Define wasting based on WFHZ, MFAZ, MUAC and Combined criteria}
\usage{
define_wasting_cases_muac(muac, edema = NULL, cases = c("gam", "sam", "mam"))

define_wasting_cases_whz(zscore, edema = NULL, cases = c("gam", "sam", "mam"))

define_wasting_cases_combined(
  zscore,
  muac,
  edema = NULL,
  cases = c("cgam", "csam", "cmam")
)

define_wasting(
  df,
  zscore = NULL,
  muac = NULL,
  edema = NULL,
  base = c("wfhz", "muac", "combined")
)
}
\arguments{
\item{muac}{A vector of class \code{integer} of MUAC values in millimeters.}

\item{edema}{A vector of class \code{character} of edema. Code should be
"y" for presence and "n" for absence of bilateral edema. Default is \code{NULL}.}

\item{cases}{A choice of the form of wasting to be defined.}

\item{zscore}{A vector of class \code{double} of WFHZ values (with 3 decimal places).}

\item{df}{A dataset object of class \code{data.frame} to use.}

\item{base}{A choice of the criterion on which the case-definition should be based.}
}
\value{
A vector of class \code{numeric} of dummy values: 1 for case and 0
for not case.
}
\description{
Define if a given observation in the dataset is wasted or not, on the basis of
WFHZ, MFAZ, MUAC and the combined criteria.
}
\details{
Use \code{define_wasting()} to add the case-definitions to data frame.
}
\examples{

## Weight-for-height based case-definition ----
x <- anthro.02 |>
define_wasting(
zscore = wfhz,
edema = edema,
base = "wfhz"
)
head(x)

## MUAC-based case-definition ----
x <- anthro.02 |>
define_wasting(
muac = muac,
edema = edema,
base = "muac"
)
head(x)

## Combined case-definition ----
x <- anthro.02 |>
define_wasting(
zscore = wfhz,
muac = muac,
edema = edema,
base = "combined"
)
head(x)

}
