% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_raters.R
\name{classify_age_sex_ratio}
\alias{classify_age_sex_ratio}
\title{Rate the p-values of the age and sex ratio test}
\usage{
classify_age_sex_ratio(p)
}
\arguments{
\item{p}{A numeric vector containing the test p-values.}
}
\value{
A character vector with the rating results.
}
\description{
Rate the p-values of the age and sex ratio test
}
