% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_raters.R
\name{classify_overall_quality}
\alias{classify_overall_quality}
\title{Rate the overall acceptability score}
\usage{
classify_overall_quality(df)
}
\arguments{
\item{df}{A dataset of class \code{data.frame} containing a vector of the overall
acceptability score as yielded from \code{\link[=compute_quality_score]{compute_quality_score()}}.}
}
\value{
A \code{data.frame} based on \code{df}. A new column \code{quality_class} for the
overall acceptability rate is created and added to \code{df}.
}
\description{
Rate the overall acceptability score into "Excellent", "Good", "Acceptable" and
"Problematic".
}
\examples{
## A sample data ----

df <- data.frame(
quality_score = 29
)

## Apply the function ----
classify_overall_quality(df)

}
