% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_raters.R
\name{classify_skew_kurt}
\alias{classify_skew_kurt}
\title{Rate the magnitude of skewness and kurtosis test results}
\usage{
classify_skew_kurt(sk)
}
\arguments{
\item{sk}{A numeric vector containing values of either skewness or kurtosis.}
}
\value{
A character vector with the rating results.
}
\description{
Rate the magnitude of skewness and kurtosis test results
}
