% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_combined.R
\name{compute_pps_based_combined_prevalence}
\alias{compute_pps_based_combined_prevalence}
\alias{compute_combined_prevalence}
\title{Compute combined prevalence of wasting}
\usage{
compute_pps_based_combined_prevalence(
  df,
  .wt = NULL,
  .edema = NULL,
  .summary_by
)

compute_combined_prevalence(df, .wt = NULL, .edema = NULL, .summary_by = NULL)
}
\arguments{
\item{df}{An already wrangled dataset of class \code{data.frame} to use. Both
wranglers (of WFHZ and of MUAC) need to be used sequentially, regardless of the
order. Note that MUAC values should be converted to millimeters after using
the MUAC wrangler.}

\item{.wt}{A vector of class \code{double} of the final survey weights. Default is
\code{NULL} assuming a self-weighted survey, as in the ENA for SMART software;
otherwise a weighted analysis is computed.}

\item{.edema}{A vector of class \code{character} of edema. Code should be
"y" for presence and "n" for absence of bilateral edema. Default is \code{NULL}.}

\item{.summary_by}{A vector of class \code{character} of the geographical areas
where the data was collected and for which the analysis should be performed.}
}
\value{
A summarised table of class \code{data.frame} for the descriptive
statistics about combined wasting.
}
\description{
The prevalence is calculated in accordance with the complex sample design
properties inherent to surveys. This includes weighting of survey data where
applicable. When either the acceptability of the standard deviation of WFHZ or
of the age ratio test is problematic, prevalence is not calculated.
}
\details{
A concept of "combined flags" is introduced in this function. It consists of
defining as flag any observation that is flagged in either \code{flag_wfhz} or
\code{flag_mfaz} vectors. A new column \code{cflags} for combined flags is created and
added to \code{df}. This ensures that all flagged observations from both WFHZ
and MFAZ data are excluded from the combined prevalence analysis.

\emph{The table below shows an overview of how \code{cflags} are defined}\tabular{ccc}{
   \strong{flag_wfhz} \tab \strong{flag_mfaz} \tab \strong{cflags} \cr
   1 \tab 0 \tab 1 \cr
   0 \tab 1 \tab 1 \cr
   0 \tab 0 \tab 0 \cr
}
}
\examples{

## When .summary_by and .wt are set to NULL ----
p <- compute_combined_prevalence(
df = anthro.02,
.wt = NULL,
.edema = edema,
.summary_by = NULL
)

print(p)

## When .wt is not set to NULL ----
x <- compute_combined_prevalence(
df = anthro.02,
.wt = "wtfactor",
.edema = edema,
.summary_by = NULL
)

print(x)

## When working on data frame with multiple survey areas ----
s <- anthro.03 |>
mw_wrangle_age(
dos = NULL,
dob = NULL,
age = age,
.decimals = 2
) |>
mw_wrangle_muac(
sex = sex,
muac = muac,
age = "age",
.recode_sex = TRUE,
.recode_muac = TRUE,
.to = "cm"
) |>
dplyr::mutate(muac = recode_muac(muac, .to = "mm")) |>
mw_wrangle_wfhz(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE) |>
compute_combined_prevalence(
.edema = edema,
.summary_by = district
)

print(s)

}
