% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{compute_age_in_months}
\alias{compute_age_in_months}
\title{Calculate child's age in months}
\usage{
compute_age_in_months(surv_date, birth_date)
}
\arguments{
\item{surv_date}{A vector of class \code{Date} for data collection date.}

\item{birth_date}{A vector of class \code{Date} for child's date of birth.}
}
\value{
A vector of class \code{double} for child's age in months with two decimal places.
Any value less than 6.0 and greater than or equal to 60.0 months will be set to \code{NA}.
}
\description{
Calculate child's age in months based on date of birth and the data collection date.
}
