% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{compute_month_to_days}
\alias{compute_month_to_days}
\title{Transform age in months to days}
\usage{
compute_month_to_days(x)
}
\arguments{
\item{x}{A numeric vector containing age values in months.}
}
\value{
A numeric vector, of the same length as the input variable, containing
age values in days.
}
\description{
Transform age in months to days
}
