% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scorers.R
\name{compute_quality_score}
\alias{compute_quality_score}
\title{Get the overall acceptability score from the acceptability classification scores}
\usage{
compute_quality_score(df, type = c("mfaz", "whz"))
}
\arguments{
\item{df}{A dataset object of class \code{data.frame} to calculate from.}

\item{type}{A choice between "wfhz" and "mfaz" for the basis on which the
calculations should be made.}
}
\value{
A \code{data.frame} based on \code{df} with a new column named \code{"quality_score"}
for the overall of acceptability (of quality) score.
}
\description{
Calculate the total amount of penalty points based on each plausibility test
result acceptability classification for WFHZ and MFAZ.
}
\examples{

## A sample data ----

df <- data.frame(
flagged_class = "Excellent",
age_ratio_class = "Good",
sex_ratio_class = "Problematic",
dps_class = "Excellent",
sd_class = "Excellent",
skew_class = "Good",
kurt_class = "Acceptable"
)

## Apply the function ----
compute_quality_score(df, type = "mfaz")

}
