% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_define_wasting.R
\name{define_wasting}
\alias{define_wasting}
\title{Define wasting}
\usage{
define_wasting(
  df,
  zscores = NULL,
  muac = NULL,
  edema = NULL,
  .by = c("zscores", "muac", "combined")
)
}
\arguments{
\item{df}{A \code{tibble} object. It must have been wrangled using this package's
wrangling functions for WFHZ or MUAC, or both (for combined) as appropriate.}

\item{zscores}{A vector of class \code{double} of WFHZ or MFAZ values.}

\item{muac}{An \code{integer} or \code{character} vector of raw MUAC values in
millimeters.}

\item{edema}{A \code{character} vector indicating edema status. Default is NULL.
Code values should be "y" for presence and "n" for absence of nutritional
edema.}

\item{.by}{A choice of the criterion by which a case is to be defined. Choose
"zscores" for WFHZ or MFAZ, "muac" for raw MUAC and "combined" for combined.
Default value is "zscores".}
}
\value{
The \code{tibble} object \code{df} with additional columns named named \code{gam},
\code{sam} and \code{mam}, each of class \code{numeric} containing coded values of either
1 (case) and 0 (not a case). If \code{.by = "combined"}, additional columns are
named \code{cgam}, \code{csam} and \code{cmam}.
}
\description{
Determine if a given observation in the data set is wasted or not, and its
respective form of wasting (global, severe or moderate) on the basis of
z-scores of weight-for-height (WFHZ), muac-for-age (MFAZ), raw MUAC
values and combined case-definition.
}
\examples{
## Case-definition by z-scores ----
z <- anthro.02 |>
  define_wasting(
    zscores = wfhz,
    muac = NULL,
    edema = edema,
    .by = "zscores"
  )
head(z)

## Case-definition by MUAC ----
m <- anthro.02 |>
  define_wasting(
    zscores = NULL,
    muac = muac,
    edema = edema,
    .by = "muac"
  )
head(m)

## Case-definition by combined ----
c <- anthro.02 |>
  define_wasting(
    zscores = wfhz,
    muac = muac,
    edema = edema,
    .by = "combined"
  )
head(c)

}
