% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_definitions.R
\name{define_wasting}
\alias{define_wasting}
\title{Define wasting}
\usage{
define_wasting(
  df,
  zscores = NULL,
  muac = NULL,
  edema = NULL,
  .by = c("zscores", "muac", "combined")
)
}
\arguments{
\item{df}{A dataset object of class \code{data.frame} to use. It must have been
wrangled using this package's wrangling functions for WFHZ or MUAC, or both
(for combined) as appropriate.}

\item{zscores}{A vector of class \code{double} of WFHZ or MFAZ values. If the class
does not match the expected type, the function will stop execution and return
an error message indicating the type of mismatch.}

\item{muac}{A vector of class \code{integer} or \code{numeric} of raw MUAC values in
millimeters. If the class does not match the expected type, the function will
stop execution and return an error message indicating the type of mismatch.}

\item{edema}{A vector of class \code{character} of edema. Default is \code{NULL}.
If the class does not match the expected type, the function will stop execution
and return an error message indicating the type of mismatch. Code values should be
"y" for presence and "n" for absence of bilateral edema. If different, the
function will stop execution and return an error indicating the issue.}

\item{.by}{A choice of the criterion by which the case-definition should done.
Choose \code{zscores} for WFHZ or MFAZ, \code{muac} for raw MUAC and \code{combined} for
combined.}
}
\value{
Three vectors named \code{gam}, \code{sam} and \code{mam}, of class \code{numeric}, same
length as inputs, containing dummy values: 1 for case and 0 for not case.
This is added to \code{df}. When \code{combined} is selected, vector's names become
\code{cgam}, \code{csam} and \code{cmam}.
}
\description{
Define if a given observation in the dataset is wasted or not, and its
respective form of wasting (global, severe or moderate) on the basis of
z-scores of weight-for-height (WFHZ), muac-for-age (MFAZ), raw MUAC values and
combined case-definition.
}
\examples{
## Case-definition by z-scores ----
z <- anthro.02 |>
  define_wasting(
    zscores = wfhz,
    muac = NULL,
    edema = edema,
    .by = "zscores"
  )
head(z)

## Case-definition by MUAC ----
m <- anthro.02 |>
  define_wasting(
    zscores = NULL,
    muac = muac,
    edema = edema,
    .by = "muac"
  )
head(m)

## Case-definition by combined ----
c <- anthro.02 |>
  define_wasting(
    zscores = wfhz,
    muac = muac,
    edema = edema,
    .by = "combined"
  )
head(c)

}
