% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipccheckr-package.R
\docType{package}
\name{ipccheckr-package}
\alias{ipccheckr}
\alias{ipccheckr-package}
\title{ipccheckr: Utilities for Analysing Children's Nutritional Status}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A streamlined and comprehensive implementation of the Standardized Monitoring and Assessment of Relief and Transition (SMART) Methodology \url{https://smartmethodology.org/} guidelines for data quality checks and prevalence estimation, with enhanced programmable process particularly when handling large multiple datasets.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nutriverse/mwana}
  \item \url{https://nutriverse.io/mwana}
  \item \url{http://nutriverse.io/mwana/}
  \item Report bugs at \url{https://github.com/nutriverse/mwana/issues}
}

}
\author{
\strong{Maintainer}: Tomás Zaba \email{tomas.zaba@outlook.com} (\href{https://orcid.org/0000-0002-7079-3574}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Ernest Guevarra (\href{https://orcid.org/0000-0002-4887-4415}{ORCID}) [copyright holder]
}

}
\keyword{internal}
