% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_helpers.R
\name{mod_call_ipcamn_checker}
\alias{mod_call_ipcamn_checker}
\title{Call mwana's IPC Acute Malnutrition minimum sample size requirement checker
into the module server}
\usage{
mod_call_ipcamn_checker(df, cluster, source, area1, area2)
}
\description{
Call mwana's IPC Acute Malnutrition minimum sample size requirement checker
into the module server
}
\keyword{internal}
