% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev-wasting-screening.R
\name{mw_estimate_prevalence_screening}
\alias{mw_estimate_prevalence_screening}
\alias{mw_estimate_prevalence_screening2}
\title{Estimate the prevalence of wasting based on MUAC for non-survey data}
\usage{
mw_estimate_prevalence_screening(df, muac, age, oedema = NULL, ...)

mw_estimate_prevalence_screening2(df, age_cat, muac, oedema = NULL, ...)
}
\arguments{
\item{df}{A \code{tibble} object produced by \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}} and
\code{\link[=mw_wrangle_age]{mw_wrangle_age()}} functions. Note that MUAC values in \code{df}
must be in millimetres unit after using \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}}. Also, \code{df}
must have a variable called \code{cluster} wherein the primary sampling unit
identifiers are stored.}

\item{muac}{A \code{numeric} or \code{integer} vector of raw MUAC values. The
measurement unit should be millimetres.}

\item{age}{A vector of class \code{double} of child's age in months.}

\item{oedema}{A \code{character} vector for presence of nutritional oedema. Code
values should be "y" for presence and "n" for absence. Default is NULL.}

\item{...}{A vector of class \code{character}, specifying the categories for which
the analysis should be summarised for. Usually geographical areas. More than
one vector can be specified.}

\item{age_cat}{A \code{character} vector of child's age in categories. Code values
should be "6-23" and "24-59".}
}
\value{
A summary \code{tibble} for the descriptive statistics about wasting based
on MUAC, with no confidence intervals.
}
\description{
It is common to estimate prevalence of wasting from non-survey data, such
as screenings or any other data derived from community-based surveillance
systems. In such situations, the analysis usually consists only in estimating
the point prevalence and the counts of positive cases, without necessarily
estimating the uncertainty. This function serves this purpose.

The quality of the data is first evaluated by calculating and rating the
standard deviation (SD) of MFAZ (in \code{mw_estimate_prevalence_screening()})
or SD of the raw MUAC values (in \code{mw_estimate_prevalence_screening2()}),
and the p-value of the age ratio test in either functions.
Thereafter, if the latter test is problematic, age-weighting approach is
applied to the prevalence estimation, to account for the over-representation
of younger children in the sample; otherwise, a non-age-weighted prevalence
is estimated. This means that even if the SD in either functions is
problematic, the prevalence is estimated, with no adjustments, and returned.
}
\details{
A typical user analysis workflow is expected to begin with data quality checks,
followed by a thorough review, and only thereafter proceed to prevalence
estimation. This sequence places the user in the strongest position to assess
whether the resulting prevalence estimates are reliable.

In \code{mw_estimate_prevalence_screening()}, outliers are identified using SMART
flagging criteria applied to MFAZ, whereas \code{mw_estimate_prevalence_screening2()}
are based on the raw MUAC values. In either functions, outliers are excluded
from the prevalence estimation.
}
\examples{
mw_estimate_prevalence_screening(
  df = anthro.02,
  muac = muac,
  age = age,
  oedema = oedema,
  province
)

## With `oedema` set to `NULL` ----
mw_estimate_prevalence_screening(
  df = anthro.02,
  muac = muac,
  age = age,
  oedema = NULL,
  province
)

## Specifying the grouping variables ----
mw_estimate_prevalence_screening(
  df = anthro.02,
  muac = muac,
  age = age,
  oedema = NULL,
  province
)


anthro.01 |>
  mw_wrangle_muac(
    sex = sex,
    .recode_sex = TRUE,
    muac = muac
  ) |>
  transform(
    age_cat = ifelse(age < 24, "6-23", "24-59")
  ) |>
  mw_estimate_prevalence_screening2(
    age_cat = age_cat,
    muac = muac,
    oedema = oedema,
    area
  )

}
\references{
SMART Initiative (no date). \emph{Updated MUAC data collection tool}. Available at:
\url{https://smartmethodology.org/survey-planning-tools/updated-muac-tool/}
}
\seealso{
\code{\link[=mw_estimate_prevalence_muac]{mw_estimate_prevalence_muac()}}, \code{\link[=mw_estimate_age_weighted_prev_muac]{mw_estimate_age_weighted_prev_muac()}},
\code{\link[=flag_outliers]{flag_outliers()}} and \code{\link[=remove_flags]{remove_flags()}}.
}
