% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_amn_check.R
\name{mw_check_ipcamn_ssreq}
\alias{mw_check_ipcamn_ssreq}
\title{Check whether IPC Acute Malnutrition (IPC AMN) sample size requirements were met}
\usage{
mw_check_ipcamn_ssreq(df, cluster, .source = c("survey", "screening", "ssite"))
}
\arguments{
\item{df}{A dataset object of class \code{data.frame} to check.}

\item{cluster}{A vector of class \code{integer} or \code{character} of unique cluster or
screening or sentinel site IDs. If a \code{character} vector, ensure that names are
correct and each name represents one location for accurate counts. If the class
does not match the above expected type, the function will stop execution and
return an error message indicating the type of mismatch.}

\item{.source}{The source of evidence. A choice between "survey" for
representative survey data at the area of analysis; "screening" for
screening data; "ssite" for community-based sentinel site data.}
}
\value{
A summary table of class \code{data.frame}, of length 3 and width 1, for
the check results. \code{n_clusters} is for the total number of unique clusters or
screening or site IDs; \code{n_obs} for the correspondent total number of children
in the dataset; and \code{meet_ipc} for whether the IPC AMN requirements were met.
}
\description{
Evidence on the prevalence of acute malnutrition used in the IPC AMN
can come from different sources: surveys, screenings or community-based
surveillance system. The IPC set minimum sample size requirements
for each source. This function helps in verifying whether the requirements
were met or not depending on the source.
}
\examples{
mw_check_ipcamn_ssreq(
  df = anthro.01,
  cluster = cluster,
  .source = "survey"
)

}
\references{
IPC Global Partners. 2021. \emph{Integrated Food Security Phase Classification}
\emph{Technical Manual Version 3.1.Evidence and Standards for Better Food Security}
\emph{and Nutrition Decisions}. Rome. Available at:
\url{https://www.ipcinfo.org/ipcinfo-website/resources/ipc-manual/en/}.
}
