% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipc_amn_check.R
\name{mw_check_ipcamn_ssreq}
\alias{mw_check_ipcamn_ssreq}
\title{Check whether sample size requirements for IPC Acute Malnutrition (IPC AMN)
analysis are met}
\usage{
mw_check_ipcamn_ssreq(
  df,
  cluster,
  .source = c("survey", "screening", "ssite"),
  .by = NULL
)
}
\arguments{
\item{df}{A \code{data.frame} object to check.}

\item{cluster}{A vector of class \code{integer} or \code{character} of unique cluster
or screening or sentinel site identifiers. If a \code{character} vector, ensure
that each unique name represents one location. If \code{cluster} is not of class
\code{integer} or \code{character}, an error message will be returned indicating the
type of mismatch.}

\item{.source}{The source of evidence. A choice between "survey" for
representative survey data at the area of analysis; "screening" for
screening data; "ssite" for community-based sentinel site data. Default value
is "survey".}

\item{.by}{A \code{character} or \code{numeric} vector of the geographical areas
or identifiers for where the data was collected and for which the analysis
should be summarised for.}
}
\value{
A summary \code{tibble} with 3 columns (when \code{.by} is NULL) or 4 columns
(when \code{.by} is not NULL), containing check results for:
\itemize{
\item \code{n_clusters} - the total number of unique clusters or
screening or site identifiers;
\item \code{n_obs} - the corresponding total number of children in the data set; and,
\item \code{meet_ipc} - whether the IPC AMN requirements were met.
}
}
\description{
Data for estimating the prevalence of acute malnutrition used in the IPC AMN
can come from different sources: surveys, screenings or community-based
surveillance systems. The IPC has set minimum sample size requirements for
each  source. This function verifies whether these requirements are met.
}
\examples{
mw_check_ipcamn_ssreq(
  df = anthro.01,
  cluster = cluster,
  .source = "survey",
  .by = area
)

}
\references{
IPC Global Partners. 2021. \emph{Integrated Food Security Phase Classification}
\emph{Technical Manual Version 3.1.Evidence and Standards for Better Food Security}
\emph{and Nutrition Decisions}. Rome. Available at:
\url{https://www.ipcinfo.org/ipcinfo-website/resources/ipc-manual/en/}.
}
