% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev-wasting-age-weighted-muac.R
\name{mw_estimate_age_weighted_prev_muac}
\alias{mw_estimate_age_weighted_prev_muac}
\title{Estimate age-weighted prevalence of wasting by MUAC}
\usage{
mw_estimate_age_weighted_prev_muac(
  df,
  muac,
  has_age = TRUE,
  age = NULL,
  age_cat = NULL,
  oedema = NULL,
  raw_muac = FALSE,
  ...
)
}
\arguments{
\item{df}{A \code{tibble} object produced by \code{mwana} data wranglers.}

\item{muac}{A \code{numeric} or \code{integer} vector of raw MUAC values. The
measurement unit should be millimeters.}

\item{has_age}{Logical. Specifies whether the input dataset provides age in
months or in categories ('6–23', '24–59'). Defaults to \code{TRUE} when age is
given in months.}

\item{age}{A vector of class \code{double} of child's age in months. Defaults to
NULL. Only use if \code{has_age = TRUE}, otherwise set it to \code{NULL}.}

\item{age_cat}{A \code{character} vector of child's age in categories. Code values
should be "6-23" and "24-59". Defaults to \code{NULL}. Only use it if \code{has_age = FALSE}.}

\item{oedema}{A \code{character} vector for presence of nutritional oedema. Code
values should be "y" for presence and "n" for absence. Defaults to NULL.}

\item{raw_muac}{Logical. Whether outliers should be excluded based on the raw
MUAC values or MFAZ. For the former, set it to \code{TRUE}, otherwise \code{FALSE}.
Defaults to MFAZ.}

\item{...}{A vector of class \code{character}, specifying the categories for which
the analysis should be summarised for. Usually geographical areas. More than
one vector can be specified.}
}
\value{
A summary \code{tibble} with wasting prevalence estimates, as given by the
SMART updated MUAC tool (see references below).
}
\description{
Estimates age‑weighted prevalence of wasting using MUAC.
Accepts age in months or in categories ('6–23', '24–59').
The default is age in months.

The prevalence is weighted as:
\deqn{( prevalence_{6-23} + 2 \times prevalence_{24-59} ) / 3}

Whilst the function is exported to users as a standalone, it is embedded into
the following major MUAC prevalence functions of this package:
\code{\link[=mw_estimate_prevalence_muac]{mw_estimate_prevalence_muac()}}, \code{\link[=mw_estimate_prevalence_screening]{mw_estimate_prevalence_screening()}}, and
\code{\link[=mw_estimate_prevalence_screening2]{mw_estimate_prevalence_screening2()}}.
}
\details{
As a standalone function, the user must check data quality before calling the function.
}
\examples{
## Example application when age is given in months ----
mw_estimate_age_weighted_prev_muac(
  df = anthro.04,
  muac = muac,
  has_age = TRUE,
  age = age,
  age_cat = FALSE,
  oedema = oedema,
  raw_muac = FALSE,
  province
)

## Example application when age is given in categories ----
anthro.04 |>
  transform(age_cat = ifelse(age < 24, "6-23", "24-59")) |>
  mw_wrangle_muac(
    muac = muac,
    .recode_muac = FALSE,
    .to = "none",
    sex = sex,
    .recode_sex = FALSE
  ) |>
  mw_estimate_age_weighted_prev_muac(
    has_age = FALSE,
    age = NULL,
    age_cat = age_cat,
    oedema = oedema,
    raw_muac = FALSE
  )

}
\references{
SMART Initiative (no date). \emph{Updated MUAC data collection tool}. Available at:
\url{https://smartmethodology.org/survey-planning-tools/updated-muac-tool/}
}
