% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev-wasting-combined.R
\name{mw_estimate_prevalence_combined}
\alias{mw_estimate_prevalence_combined}
\title{Estimate the prevalence of combined wasting}
\usage{
mw_estimate_prevalence_combined(df, wt = NULL, oedema = NULL, ...)
}
\arguments{
\item{df}{A \code{tibble} object produced by sequential application of the
\code{\link[=mw_wrangle_wfhz]{mw_wrangle_wfhz()}} and \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}}. Note that MUAC values in \code{df}
must be in millimetres unit after using \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}}. In addition, \code{df}
must have a variable called \code{cluster}, which contains the primary sampling
unit identifiers.}

\item{wt}{A vector of class \code{double} of the survey sampling weights. Default
is NULL which assumes a self-weighted survey as is the case for a survey
sample selected proportional to population size (i.e., SMART survey sample).
Otherwise, a weighted analysis is implemented.}

\item{oedema}{A \code{character} vector for presence of nutritional oedema coded as
"y" for presence of nutritional oedema and "n" for absence of nutritional
oedema. Default is NULL.}

\item{...}{A vector of class \code{character}, specifying the categories for which
the analysis should be summarised for. Usually geographical areas. More than
one vector can be specified.}
}
\value{
A summary \code{tibble} for the descriptive statistics about combined
wasting.
}
\description{
Estimate the prevalence of wasting based on the combined case-definition of
weight-for-height z-scores (WFHZ), MUAC and/or oedema. The function allows
users to estimate prevalence in accordance with complex-sample design
properties such as accounting for survey sample weights when needed or
applicable.

The data quality is first assessed by calculating and rating the standard
deviation (SD) of WFHZ. Then it calculates the observed proportion of children
aged 24–59 months out of all children in the dataset. Next, it estimates the
p-value for the difference between this observed proportion and the expected
(0.66), and rates the result.

Prevalence is estimated only when the WFHZ SD is not problematic and the age
ratio test is not problematic, or — if the age ratio test is problematic — the
proportion of children aged 24–59 months is greater than or equal to 0.66.
}
\details{
A concept of \emph{combined flags} is introduced in this function. Any observation
that is flagged for either \code{flag_wfhz} or \code{flag_mfaz} is flagged under a new
variable named \code{cflags} added to \code{df}. This ensures that all flagged
observations from both WFHZ and MFAZ data are excluded from the prevalence
analysis.\tabular{ccc}{
   \strong{flag_wfhz} \tab \strong{flag_mfaz} \tab \strong{cflags} \cr
   1 \tab 0 \tab 1 \cr
   0 \tab 1 \tab 1 \cr
   0 \tab 0 \tab 0 \cr
}
}
\examples{
## When wt are set to NULL ----
mw_estimate_prevalence_combined(
  df = anthro.02,
  wt = NULL,
  oedema = oedema
)

## When `wt` is not set to NULL ----
mw_estimate_prevalence_combined(
  df = anthro.02,
  wt = wtfactor,
  oedema = oedema
)

}
