% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_combined.R
\name{mw_estimate_prevalence_combined}
\alias{mw_estimate_prevalence_combined}
\title{Estimate the prevalence of combined wasting}
\usage{
mw_estimate_prevalence_combined(df, wt = NULL, edema = NULL, .by = NULL)
}
\arguments{
\item{df}{A dataset object of class \code{data.frame} to use. This must have been
wrangled using this package's wrangling functions for both WFHZ and MUAC data
sequentially. The order does not matter. Note that MUAC values should be
converted to millimeters after using the MUAC wrangler. If this is not done,
the function will stop execution and return an error message. Moreover, the
function uses a variable called \code{cluster} where the primary sampling unit IDs
are stored. Make sure to rename your cluster ID variable to \code{cluster}, otherwise
the function will error and terminate the execution.}

\item{wt}{A vector of class \code{double} of the final survey weights. Default is
\code{NULL} assuming a self-weighted survey, as in the ENA for SMART software;
otherwise a weighted analysis is computed.}

\item{edema}{A vector of class \code{character} of edema. Code will be
"y" for presence and "n" for absence of bilateral edema. Default is \code{NULL}.}

\item{.by}{A vector of class \code{character} or \code{numeric} of the geographical areas
or respective IDs for where the data was collected and for which the analysis
should be summarised at.}
}
\value{
A summarised table of class \code{data.frame} for the descriptive
statistics about combined wasting.
}
\description{
Estimate the prevalence of wasting based on the combined case-definition of
weight-for-height z-scores (WFHZ), MUAC and/or edema. The function allows users to
get the prevalence estimates in accordance with the complex sample
design properties; this includes applying survey weights when needed or applicable.
Before estimating, the function evaluates the quality of data by calculating
and rating the standard deviation of WFHZ and MFAZ, as well as the p-value of
the age ratio test.
Prevalence will be calculated only when the rating of all test is as not
problematic concurrently. If either of them is problematic, it cancels out
the analysis and \code{NA}s get thrown.

Outliers are detected in both WFHZ and in MUAC data set (through z-scores)
based on SMART flags get excluded prior being piped into the actual prevalence
analysis workflow.
}
\details{
A concept of "combined flags" is introduced in this function. It consists of
defining as flag any observation that is flagged in either \code{flag_wfhz} or
\code{flag_mfaz} vectors. A new column \code{cflags} for combined flags is created and
added to \code{df}. This ensures that all flagged observations from both WFHZ
and MFAZ data are excluded from the prevalence analysis.

\emph{A glimpse on how \code{cflags} are defined:}\tabular{ccc}{
   \strong{flag_wfhz} \tab \strong{flag_mfaz} \tab \strong{cflags} \cr
   1 \tab 0 \tab 1 \cr
   0 \tab 1 \tab 1 \cr
   0 \tab 0 \tab 0 \cr
}
}
\examples{
## When .by and wt are set to NULL ----
mw_estimate_prevalence_combined(
  df = anthro.02,
  wt = NULL,
  edema = edema,
  .by = NULL
)

## When wt is not set to NULL ----
mw_estimate_prevalence_combined(
  df = anthro.02,
  wt = wtfactor,
  edema = edema,
  .by = NULL
)

}
