% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_combined.R
\name{mw_estimate_prevalence_combined}
\alias{mw_estimate_prevalence_combined}
\title{Estimate the prevalence of combined wasting}
\usage{
mw_estimate_prevalence_combined(df, wt = NULL, edema = NULL, ...)
}
\arguments{
\item{df}{A \code{tibble} object produced by sequential application of the
\code{\link[=mw_wrangle_wfhz]{mw_wrangle_wfhz()}} and \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}}. Note that MUAC values in \code{df}
must be in millimeters unit after using \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}}. Also, \code{df}
must have a variable called \code{cluster} which contains the primary sampling
unit identifiers.}

\item{wt}{A vector of class \code{double} of the survey sampling weights. Default
is NULL which assumes a self-weighted survey as is the case for a survey
sample selected proportional to population size (i.e., SMART survey sample).
Otherwise, a weighted analysis is implemented.}

\item{edema}{A \code{character} vector for presence of nutritional edema coded as
"y" for presence of nutritional edema and "n" for absence of nutritional
edema. Default is NULL.}

\item{...}{A vector of class \code{character}, specifying the categories for which
the analysis should be summarised for. Usually geographical areas. More than
one vector can be specified.}
}
\value{
A summary \code{tibble} for the descriptive statistics about combined
wasting.
}
\description{
Estimate the prevalence of wasting based on the combined case-definition of
weight-for-height z-scores (WFHZ), MUAC and/or edema. The function allows
users to estimate prevalence in accordance with complex sample design
properties such as accounting for survey sample weights when needed or
applicable. The quality of the data is first evaluated by calculating and
rating the standard deviation of WFHZ and MFAZ and the p-value of the age
ratio test. Prevalence is calculated only when all tests are rated as not
problematic. If any of the tests rate as problematic, no estimation is done
and an NA value is returned. Outliers are detected in both WFHZ and MFAZ
datasets based on SMART flagging criteria. Identified outliers are then
excluded before prevalence estimation is performed.
}
\details{
A concept of \emph{combined flags} is introduced in this function. Any observation
that is flagged for either \code{flag_wfhz} or \code{flag_mfaz} is flagged under a new
variable named \code{cflags} added to \code{df}. This ensures that all flagged
observations from both WFHZ and MFAZ data are excluded from the prevalence
analysis.\tabular{ccc}{
   \strong{flag_wfhz} \tab \strong{flag_mfaz} \tab \strong{cflags} \cr
   1 \tab 0 \tab 1 \cr
   0 \tab 1 \tab 1 \cr
   0 \tab 0 \tab 0 \cr
}
}
\examples{
## When wt are set to NULL ----
mw_estimate_prevalence_combined(
  df = anthro.02,
  wt = NULL,
  edema = edema
)

## When `wt` is not set to NULL ----
mw_estimate_prevalence_combined(
  df = anthro.02,
  wt = wtfactor,
  edema = edema
)

}
