% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_mfaz.R
\name{mw_estimate_prevalence_mfaz}
\alias{mw_estimate_prevalence_mfaz}
\title{Estimate the prevalence of wasting based on z-scores of muac-for-age (MFAZ)}
\usage{
mw_estimate_prevalence_mfaz(df, wt = NULL, edema = NULL, ...)
}
\arguments{
\item{df}{A \code{data.frame} object that has been produced by the
\code{\link[=mw_wrangle_age]{mw_wrangle_age()}} and \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}} functions. The \code{df} should have a
variable named \code{cluster} for the primary sampling unit identifiers.}

\item{wt}{A vector of class \code{double} of the survey sampling weights. Default
is NULL which assumes a self-weighted survey as is the case for a survey
sample selected proportional to population size (i.e., SMART survey sample).
Otherwise, a weighted analysis is implemented.}

\item{edema}{A \code{character} vector for presence of nutritional edema coded as
"y" for presence of nutritional edema and "n" for absence of nutritional
edema. Default is NULL.}

\item{...}{A vector of class \code{character}, specifying the categories for which
the analysis should be summarised for. Usually geographical areas. More than
one vector can be specified.}
}
\value{
A summary \code{tibble} for the descriptive statistics about wasting.
}
\description{
Calculate the prevalence estimates of wasting based on z-scores of
MUAC-for-age and/or bilateral edema. The function allows users to estimate
prevalence in accordance with complex sample design properties such as
accounting for survey sample weights when needed or applicable. The quality
of the data is first evaluated by calculating and rating the standard
deviation of MFAZ. Standard approach to prevalence estimation is calculated
only when the standard deviation of MFAZ is rated as not problematic. If
the standard deviation is problematic, prevalence is estimated using the
PROBIT estimator. Outliers are detected based on SMART flagging criteria.
Identified outliers are then excluded before prevalence estimation is
performed.
}
\examples{
## Without grouping variables ----
mw_estimate_prevalence_mfaz(
  df = anthro.04,
  wt = NULL,
  edema = edema
)

## With grouping variables ----
mw_estimate_prevalence_mfaz(
  df = anthro.04,
  wt = NULL,
  edema = edema,
  province
)

}
