% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_mfaz.R
\name{mw_estimate_prevalence_mfaz}
\alias{mw_estimate_prevalence_mfaz}
\title{Estimate the prevalence of wasting based on z-scores of muac-for-age (MFAZ)}
\usage{
mw_estimate_prevalence_mfaz(df, wt = NULL, edema = NULL, .by = NULL)
}
\arguments{
\item{df}{A dataset object of class \code{data.frame} to use. This must have been
wrangled using this package's wrangling function for MUAC data. The function
uses a variable name called \code{cluster} where the primary sampling unit IDs
are stored. Make sure to rename your cluster ID variable to \code{cluster}, otherwise
the function will error and terminate the execution.}

\item{wt}{A vector of class \code{double} of the final survey weights. Default is
\code{NULL} assuming a self weighted survey, as in the ENA for SMART software;
otherwise, when a vector of weights if supplied, weighted analysis is done.}

\item{edema}{A vector of class \code{character} of edema. Code should be
"y" for presence and "n" for absence of bilateral edema. Default is \code{NULL}.}

\item{.by}{A vector of class \code{character} or \code{numeric} of the geographical areas
or respective IDs for where the data was collected and for which the analysis
should be summarized at.}
}
\value{
A summarized table of class \code{data.frame} of the descriptive
statistics about wasting.
}
\description{
Calculate the prevalence estimates of wasting based on z-scores of
muac-for-age and/or bilateral edema. The function allows users to
get the prevalence estimates calculated in accordance with the complex sample
design properties; this includes applying survey weights when needed or applicable.

Before estimating, the function evaluates the quality of data by calculating
and rating the standard deviation of z-scores of MFAZ. If rated as problematic,
the prevalence is estimated based on the PROBIT method.

Outliers are detected based on SMART flags and get excluded prior prevalence analysis.
}
\examples{
## When .by = NULL ----
mw_estimate_prevalence_mfaz(
  df = anthro.04,
  wt = NULL,
  edema = edema,
  .by = NULL
)

## When .by is not set to NULL ----
mw_estimate_prevalence_mfaz(
  df = anthro.04,
  wt = NULL,
  edema = edema,
  .by = province
)

}
