% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev-wasting-muac.R
\name{mw_estimate_prevalence_muac}
\alias{mw_estimate_prevalence_muac}
\title{Estimate the prevalence of wasting based on MUAC for survey data}
\usage{
mw_estimate_prevalence_muac(df, age, wt = NULL, oedema = NULL, ...)
}
\arguments{
\item{df}{A \code{tibble} object produced by \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}} and
\code{\link[=mw_wrangle_age]{mw_wrangle_age()}} functions. Note that MUAC values in \code{df}
must be in millimetres after using \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}}. Also, \code{df}
must have a variable called \code{cluster} wherein the primary sampling unit
identifiers are stored.}

\item{age}{A vector of class \code{double} of child's age in months.}

\item{wt}{A vector of class \code{double} of the survey sampling weights. Default
is NULL, which assumes a self-weighted survey, the case of SMART surveys.
Otherwise, a weighted analysis is implemented.}

\item{oedema}{A \code{character} vector for presence of nutritional oedema Code
values should be "y" for presence and "n" for absence. Default is NULL.}

\item{...}{A vector of class \code{character}, specifying the categories for which
the analysis should be summarised for. Usually geographical areas. More than
one vector can be specified.}
}
\value{
A summary \code{tibble} for the descriptive statistics about wasting based
on MUAC, with confidence intervals.
}
\description{
Estimate the prevalence of wasting based on MUAC and/or nutritional oedema.
The function allows users to estimate prevalence in accordance with complex
sample design properties, such as accounting for survey sample weights when
needed or applicable.

The quality of the data is first evaluated by calculating and rating the
standard deviation (SD) of MFAZ and the p-value of the age ratio test.
Thereafter, if the latter test is problematic, age-weighting approach is
applied to prevalence estimation, to account for the over-representation
of younger children in the sample; otherwise, a non-age-weighted prevalence
is estimated. This means that even if the SD of MFAZ is problematic, the
prevalence is estimated, with no adjustments, and returned.
}
\details{
A typical user analysis workflow is expected to begin with data quality checks,
followed by a thorough review, and only thereafter proceed to prevalence
estimation. This sequence places the user in the strongest position to assess
whether the resulting prevalence estimates are reliable.

Outliers are identified using SMART flagging criteria applied to MFAZ, and are
excluded from the prevalence estimation.
}
\examples{
## Ungrouped analysis ----
mw_estimate_prevalence_muac(
  df = anthro.04,
age = age,
  wt = NULL,
  oedema = oedema
)

## Grouped analysis ----
mw_estimate_prevalence_muac(
  df = anthro.04,
age = age,
  wt = NULL,
  oedema = oedema,
  province
)

}
\references{
SMART Initiative (no date). \emph{Updated MUAC data collection tool}. Available at:
\url{https://smartmethodology.org/survey-planning-tools/updated-muac-tool/}
}
\seealso{
\code{\link[=mw_estimate_age_weighted_prev_muac]{mw_estimate_age_weighted_prev_muac()}} \code{\link[=mw_estimate_prevalence_mfaz]{mw_estimate_prevalence_mfaz()}}
\code{\link[=mw_estimate_prevalence_screening]{mw_estimate_prevalence_screening()}}
}
