% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_screening.R
\name{mw_estimate_prevalence_screening}
\alias{mw_estimate_prevalence_screening}
\title{Estimate the prevalence of wasting based on MUAC for non-survey data}
\usage{
mw_estimate_prevalence_screening(df, muac, edema = NULL, .by = NULL)
}
\arguments{
\item{df}{A \code{tibble} object produced by \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}} and
\code{\link[=mw_wrangle_age]{mw_wrangle_age()}} functions. Note that MUAC values in \code{df}
must be in millimeters unit after using \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}}. Also, \code{df}
must have a variable called \code{cluster} which contains the primary sampling
unit identifiers.}

\item{muac}{A \code{numeric} or \code{integer} vector of raw MUAC values. The
measurement unit of the values should be millimeters.}

\item{edema}{A \code{character} vector for presence of nutritional edema coded as
"y" for presence of nutritional edema and "n" for absence of nutritional
edema. Default is NULL.}

\item{.by}{A \code{character} or \code{numeric} vector of the geographical areas
or identifiers for where the data was collected and for which the analysis
should be summarised for.}
}
\value{
A summary \code{tibble} for the descriptive statistics about combined
wasting.
}
\description{
It is common to estimate prevalence of wasting from non survey data, such
as screenings or any other community-based surveillance systems. In such
situations, the analysis usually consists only in estimating the point
prevalence and the counts of positive cases, without necessarily estimating
the uncertainty. This function serves this use.

The quality of the data is first evaluated by calculating and rating the
standard deviation of MFAZ and the p-value of the age ratio test. Prevalence
is calculated only when the standard deviation of MFAZ is not problematic. If
both standard deviation of MFAZ and p-value of age ratio test is not
problematic, straightforward prevalence estimation is performed. If standard
deviation of MFAZ is not problematic but p-value of age ratio test is
problematic, age-weighting is applied to prevalence estimation to account for
the over-representation of younger children in the sample. If standard
deviation of MFAZ is problematic, no estimation is done and an NA value is
returned. Outliers are detected based on SMART flagging criteria for MFAZ.
Identified outliers are then excluded before prevalence  estimation is
performed.
}
\examples{
mw_estimate_prevalence_screening(
  df = anthro.02,
  muac = muac,
  edema = edema,
  .by = province
)

## With `edema` set to `NULL` ----
mw_estimate_prevalence_screening(
  df = anthro.02,
  muac = muac,
  edema = NULL,
  .by = province
)

## With `.by` set to `NULL` ----
mw_estimate_prevalence_screening(
  df = anthro.02,
  muac = muac,
  edema = NULL,
  .by = NULL
)

}
\references{
SMART Initiative (no date). \emph{Updated MUAC data collection tool}. Available at:
\url{https://smartmethodology.org/survey-planning-tools/updated-muac-tool/}
}
\seealso{
\code{\link[=mw_estimate_prevalence_muac]{mw_estimate_prevalence_muac()}} \code{\link[=mw_estimate_smart_age_wt]{mw_estimate_smart_age_wt()}}
}
