% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_screening.R
\name{mw_estimate_prevalence_screening}
\alias{mw_estimate_prevalence_screening}
\title{Estimate the prevalence of wasting based on MUAC for non survey data}
\usage{
mw_estimate_prevalence_screening(df, muac, edema = NULL, .by = NULL)
}
\arguments{
\item{df}{A data set object of class \code{data.frame} to use. This must have been
wrangled using this package's wrangling function for MUAC data. Make sure
MUAC values are converted to millimeters after using the wrangler.
If this is not done, the function will stop execution and return an error message
with the issue.}

\item{muac}{A vector of raw MUAC values of class \code{numeric} or \code{integer}.
The measurement unit of the values should be millimeters. If any or all values
are in a different unit than the expected, the function will stop execution and
return an error message indicating the issue.}

\item{edema}{A vector of class \code{character} of edema. Code should be
"y" for presence and "n" for absence of bilateral edema. Default is \code{NULL}.
If class, as well as, code values are different than expected, the function
will stop the execution and return an error message indicating the issue.}

\item{.by}{A vector of class \code{character} or \code{numeric} of the geographical areas
or respective IDs for where the data was collected and for which the analysis
should be summarized at.}
}
\description{
It is common to estimate prevalence of wasting from non survey data, such
as screenings or any other community-based surveillance systems. In such
situations, the analysis usually consists only in estimating the point prevalence
and the counts of positive cases, without necessarily estimating the
uncertainty. This is the job of this function.

Prior estimating, it evaluates the quality of data by calculating and rating the
standard deviation of z-scores of muac-for-age (MFAZ) and the age ratio test
p-value, and then sets an analysis path that best fits the data. Paths vary between
weighted, unweighted analysis or thrown of \code{NA}s. Weighted analysis refers to the
age-weighting approach used in the SMART MUAC Tool to fix for the likely
overestimation of wasting when there are excess of younger children in the
data set. \code{NA}s get thrown when all checks are concurrently rated as problematic.
}
\examples{
mw_estimate_prevalence_screening(
df = anthro.02,
muac = muac,
edema = edema,
.by = province
)

## With `edema` set to `NULL` ----
mw_estimate_prevalence_screening(
df = anthro.02,
muac = muac,
edema = NULL,
.by = province
)

## With `.by` set to `NULL` ----
mw_estimate_prevalence_screening(
df = anthro.02,
muac = muac,
edema = NULL,
.by = NULL
)

}
\references{
SMART Initiative (no date). \emph{Updated MUAC data collection tool}. Available at:
\url{https://smartmethodology.org/survey-planning-tools/updated-muac-tool/}
}
\seealso{
\code{\link[=mw_estimate_prevalence_muac]{mw_estimate_prevalence_muac()}} \code{\link[=mw_estimate_smart_age_wt]{mw_estimate_smart_age_wt()}}
}
