% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_wfhz.R
\name{mw_estimate_prevalence_wfhz}
\alias{mw_estimate_prevalence_wfhz}
\title{Estimate the prevalence of wasting based on z-scores of weight-for-height (WFHZ)}
\usage{
mw_estimate_prevalence_wfhz(df, wt = NULL, edema = NULL, .by = NULL)
}
\arguments{
\item{df}{A dataset object of class \code{data.frame} to use. This must have been
wrangled using this package's wrangling function for WFHZ data. The function
uses a variable name called \code{cluster} where the primary sampling unit IDs
are stored. Make sure to rename your cluster ID variable to \code{cluster}, otherwise
the function will error and terminate the execution.}

\item{wt}{A vector of class \code{double} of the final survey weights. Default is
\code{NULL} assuming a self weighted survey, as in the ENA for SMART software;
otherwise, when a vector of weights if supplied, weighted analysis is done.}

\item{edema}{A vector of class \code{character} of edema. Code should be
"y" for presence and "n" for absence of bilateral edema. Default is \code{NULL}.}

\item{.by}{A vector of class \code{character} or \code{numeric} of the geographical areas
or respective IDs for where the data was collected and for which the analysis
should be summarised at.}
}
\value{
A summarised table of class \code{data.frame} of the descriptive
statistics about wasting.
}
\description{
Calculate the prevalence estimates of wasting based on z-scores of
weight-for-height and/or bilateral edema. The function allows users to
get the prevalence estimates calculated in accordance with the complex sample
design properties; this includes applying survey weights when needed or applicable.
When the standard deviation of WFHZ is rated as problematic, the prevalence is
estimated based on the PROBIT method. This is as in the SMART Methodology.
}
\examples{
## When .by = NULL ----
### Start off by wrangling the data ----
data <- mw_wrangle_wfhz(
  df = anthro.03,
  sex = sex,
  weight = weight,
  height = height,
  .recode_sex = TRUE
)

### Now run the prevalence function ----
mw_estimate_prevalence_wfhz(
  df = data,
  wt = NULL,
  edema = edema,
  .by = NULL
)

## Now when .by is not set to NULL ----
mw_estimate_prevalence_wfhz(
  df = data,
  wt = NULL,
  edema = edema,
  .by = district
)

## When a weighted analysis is needed ----
mw_estimate_prevalence_wfhz(
  df = anthro.02,
  wt = wtfactor,
  edema = edema,
  .by = province
)

}
