% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_check_mfaz.R
\name{mw_neat_output_mfaz}
\alias{mw_neat_output_mfaz}
\title{Clean and format the output tibble returned from the MUAC-for-age z-score
plausibility check}
\usage{
mw_neat_output_mfaz(df)
}
\arguments{
\item{df}{An \code{data.frame} object returned by \code{\link[=mw_plausibility_check_mfaz]{mw_plausibility_check_mfaz()}}
containing the summarized results to be formatted.}
}
\value{
A \code{data.frame} object of the same length and width as \code{df}, with column
names and values formatted as appropriate.
}
\description{
Converts scientific notations to standard notations, rounds off values, and
renames columns to meaningful names.
}
\examples{
## First wrangle age data ----
data <- mw_wrangle_age(
  df = anthro.01,
  dos = dos,
  dob = dob,
  age = age,
  .decimals = 2
)

## Then wrangle MUAC data ----
data_mfaz <- mw_wrangle_muac(
  df = data,
  sex = sex,
  age = age,
  muac = muac,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  .to = "cm"
)

## Then run plausibility check ----
pl <- mw_plausibility_check_mfaz(
  df = data_mfaz,
  flags = flag_mfaz,
  sex = sex,
  muac = muac,
  age = age,
  area
)

## Now neat the output table ----
mw_neat_output_mfaz(df = pl)

}
