% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_check_muac.R
\name{mw_neat_output_muac}
\alias{mw_neat_output_muac}
\title{Clean and format the output tibble returned from the MUAC plausibility check}
\usage{
mw_neat_output_muac(df)
}
\arguments{
\item{df}{A \code{tibble} object returned by the \code{\link[=mw_plausibility_check_muac]{mw_plausibility_check_muac()}}
function containing the summarized results to be formatted.}

\item{.by}{A \code{character} or \code{numeric} vector of the geographical areas for
where the data was collected and for which the analysis should be summarised
for.}
}
\value{
A \code{data.frame} object of the same length and width as \code{df}, with column names
and values formatted for clarity and readability.
}
\description{
Converts scientific notations to standard notations, rounds off values, and
renames columns to meaningful names.
}
\examples{
## First wrangle MUAC data ----
df_muac <- mw_wrangle_muac(
  df = anthro.01,
  sex = sex,
  muac = muac,
  age = NULL,
  .recode_sex = TRUE,
  .recode_muac = FALSE,
  .to = "none"
)

## Then run the plausibility check ----
pl_muac <- mw_plausibility_check_muac(
  df = df_muac,
  flags = flag_muac,
  sex = sex,
  muac = muac
)

## Neat the output table ----

mw_neat_output_muac(df = pl_muac)

}
