% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_check_wfhz.R
\name{mw_neat_output_wfhz}
\alias{mw_neat_output_wfhz}
\title{Clean and format the output tibble returned from the WFHZ plausibility check}
\usage{
mw_neat_output_wfhz(df, .by = NULL)
}
\arguments{
\item{df}{An \code{tibble} object returned by the \code{\link[=mw_plausibility_check_wfhz]{mw_plausibility_check_wfhz()}}
containing the summarized results to be formatted.}

\item{.by}{A \code{character} or \code{numeric} vector of the geographical areas for
where the data was collected and for which the analysis should be summarised
for.}
}
\value{
A \code{tibble} object of the same length and width as \code{df}, with column names and
values formatted for clarity and readability.
}
\description{
Converts scientific notations to standard notations, rounds off values, and
renames columns to meaningful names.
}
\examples{
## First wrangle age data ----
data <- mw_wrangle_age(
  df = anthro.01,
  dos = dos,
  dob = dob,
  age = age,
  .decimals = 2
)

## Then wrangle WFHZ data ----
data_wfhz <- mw_wrangle_wfhz(
  df = data,
  sex = sex,
  weight = weight,
  height = height,
  .recode_sex = TRUE
)

## Now run the plausibility check ----
pl <- mw_plausibility_check_wfhz(
  df = data_wfhz,
  sex = sex,
  age = age,
  weight = weight,
  height = height,
  flags = flag_wfhz,
  .by = area
)

## Now neat the output table ----
mw_neat_output_wfhz(df = pl, .by = area)

}
