% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausibility_check_wfhz.R
\name{mw_plausibility_check_wfhz}
\alias{mw_plausibility_check_wfhz}
\title{Check the plausibility and acceptability of weight-for-height z-score (WFHZ)
data}
\usage{
mw_plausibility_check_wfhz(df, sex, age, weight, height, flags)
}
\arguments{
\item{df}{A \code{tibble} object to check.}

\item{sex}{A \code{numeric} vector for child's sex with 1 = males and 2 = females.}

\item{age}{A vector of class \code{double} of child's age in months.}

\item{weight}{A vector of class \code{double} of child's weight in kilograms.}

\item{height}{A vector of class \code{double} of child's height in centimeters.}

\item{flags}{A \code{numeric} vector of flagged records.}
}
\value{
A single row summary \code{tibble} with 19 columns for the plausibility check
results and their respective acceptability rates.
}
\description{
Check the overall plausibility and acceptability of WFHZ data through a
structured test suite encompassing checks for sampling and
measurement-related biases in the dataset. The test suite, including the
criteria and corresponding rating of acceptability, follows the standards in
the SMART plausibility check.

The function works on a data frame returned by this package's wrangling
functions for age and for WFHZ data.
}
\examples{
## First wrangle age data ----
data <- mw_wrangle_age(
  df = anthro.01,
  dos = dos,
  dob = dob,
  age = age,
  .decimals = 2
)

## Then wrangle WFHZ data ----
data_wfhz <- mw_wrangle_wfhz(
  df = data,
  sex = sex,
  weight = weight,
  height = height,
  .recode_sex = TRUE
)

## Now run the plausibility check ----
mw_plausibility_check_wfhz(
  df = data_wfhz,
  sex = sex,
  age = age,
  weight = weight,
  height = height,
  flags = flag_wfhz
)


}
\references{
SMART Initiative (2017). \emph{Standardized Monitoring and Assessment for Relief
and Transition}. Manual 2.0. Available at: \url{https://smartmethodology.org}.
}
\seealso{
\code{\link[=mw_plausibility_check_mfaz]{mw_plausibility_check_mfaz()}} \code{\link[=mw_plausibility_check_muac]{mw_plausibility_check_muac()}}
\code{\link[=mw_wrangle_age]{mw_wrangle_age()}}
}
