% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stattests.R
\name{mw_stattest_ageratio}
\alias{mw_stattest_ageratio}
\title{Test for statistical difference between the proportion of children aged 24 to
59 months old over those aged 6 to 23 months old}
\usage{
mw_stattest_ageratio(age, .expectedP = 0.66)
}
\arguments{
\item{age}{A double vector of age in months.}

\item{.expectedP}{The expected proportion of children aged 24 to 59 months
old over those aged 6 to 23 months old. This is estimated to be 0.66 as in the
\href{https://smartmethodology.org/survey-planning-tools/updated-muac-tool/}{SMART MUAC tool}.}
}
\value{
A vector of class \code{list} of three statistics: \code{p} for p-value of the
statistical difference between the observed and the expected proportion of
children aged 24 to 59 months old over those aged 6 to 23 months old;
\code{observedR} and \code{observedP} for the observed ratio and proportion respectively.

@details
This function should be used specifically for assessing MUAC data. For
age ratio tests of children aged 6 to 29 months old over 30 to 59 months old, as
performed in the SMART plausibility check, use \code{\link[nipnTK:ageRatioTest]{nipnTK::ageRatioTest()}} instead.
}
\description{
Calculate the observed age ratio of children aged 24 to 59 months old over
those aged 6 to 23 months old and test if there is a statistical difference
between the observed and the expected.
}
\examples{

## An example of application using `anthro.02` dataset ----
mw_stattest_ageratio(
age = anthro.02$age,
.expectedP = 0.66
)

}
