% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_age.R
\name{mw_wrangle_age}
\alias{mw_wrangle_age}
\title{Wrangle child's age}
\usage{
mw_wrangle_age(df, dos = NULL, dob = NULL, age, .decimals = 2)
}
\arguments{
\item{df}{A dataset of class \code{data.frame} to wrangle age from.}

\item{dos}{A vector of class \code{Date} for date of data collection from the
\code{df}. Default is \code{NULL}.}

\item{dob}{A vector of class \code{Date} for child's date of birth from the \code{df}.
Default is \code{NULL}.}

\item{age}{A vector of class \code{numeric} of child's age in months. In most
cases this will be estimated using local event calendars; in some other
cases it can be a mix of the former and the one based on the child's
date of birth and the date of data collection.}

\item{.decimals}{The number of decimals places to which the age should be rounded.
Default is 2.}
}
\value{
A \code{data.frame} based on \code{df}. The variable \code{age} will be automatically
filled in each row where age value was missing and both the child's
date of birth and the date of data collection are available. Rows where \code{age}
is less than 6.0 and greater than or equal to 60.0 months old will be set to \code{NA}.
Additionally, a new variable for \code{df} named \code{age_days}, of class \code{double}, will
be created.
}
\description{
Wrangle child's age for downstream analysis. This includes calculating age
in months based on the date of data collection and the child's date of birth and
setting to \code{NA} the age values that are less than 6.0 and greater than or equal
to 60.0 months old.
}
\examples{

## A sample data ----
df <- data.frame(
  surv_date = as.Date(c(
    "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01"
  )),
  birth_date = as.Date(c(
    "2019-01-01", NA, "2018-03-20", "2019-11-05", "2021-04-25"
  )),
  age = c(NA, 36, NA, NA, NA)
)

## Apply the function ----
df |>
  mw_wrangle_age(
    dos = surv_date,
    dob = birth_date,
    age = age,
    .decimals = 3
  )

}
