% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_muac.R
\name{mw_wrangle_muac}
\alias{mw_wrangle_muac}
\title{Wrangle MUAC data}
\usage{
mw_wrangle_muac(
  df,
  sex,
  muac,
  age = NULL,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  .to = c("cm", "mm", "none"),
  .decimals = 3
)
}
\arguments{
\item{df}{A dataset of class \code{data.frame} to wrangle data from.}

\item{sex}{A \code{numeric} or \code{character} vector of child's sex. Code values should
be 1 or "m" for boy and 2 or "f" for girl.}

\item{muac}{A vector of class \code{numeric} of child's age in months.}

\item{age}{A vector of class \code{numeric} of child's age in months.}

\item{.recode_sex}{Logical. Set to \code{TRUE} if the values for \code{sex} are not coded
as 1 (for males) or 2 (for females). Otherwise, set to \code{FALSE} (default).}

\item{.recode_muac}{Logical. Set to \code{TRUE} if the raw MUAC values should be
converted to either centimeters or millimeters. Otherwise, set to \code{FALSE}
(default)}

\item{.to}{A choice of the unit to which the MUAC values should be converted.
"cm" for centimeters, "mm" for millimeters and "none" to leave as it is.}

\item{.decimals}{The number of decimals places the z-scores should have.
Default is 3.}
}
\value{
A data frame based on \code{df}. New variables named \code{mfaz} and
\code{flag_mfaz}, of child's MUAC-for-age z-scores and flags will be created. For
MUAC, when age is not supplied only \code{flag_muac} variable is created.
This refers to flags based on the raw MUAC values as recommended by
\href{https://doi.org/10.1111/mcn.13478}{Bilukha, O., & Kianian, B. (2023).}.
}
\description{
This function performs data wrangling by calculating the MUAC-for-age z-scores,
followed by the detection and flagging of outliers based on the SMART flagging
criteria. When age is not supplied, z-scores do not get computed. Instead,
outlier detection and flagging is based on the raw MUAC values.
}
\details{
A fixed flagging criterion is used for the raw MUAC values. This is as
recommended by
\href{https://doi.org/10.1111/mcn.13478}{Bilukha, O., & Kianian, B. (2023).}
}
\examples{

## When age is available ----
anthro.02 |>
  mw_wrangle_age(
    dos = NULL,
    dob = NULL,
    age = age,
    .decimals = 2
  ) |>
  mw_wrangle_muac(
    sex = sex,
    age = age,
    muac = muac,
    .recode_sex = TRUE,
    .recode_muac = TRUE,
    .to = "cm",
    .decimals = 3
  )

## When age is not available ----
anthro.02 |>
  mw_wrangle_muac(
    sex = sex,
    age = NULL,
    muac = muac,
    .recode_sex = TRUE,
    .recode_muac = TRUE,
    .to = "cm",
    .decimals = 3
  )

}
