% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_muac.R
\name{mw_wrangle_muac}
\alias{mw_wrangle_muac}
\title{Wrangle MUAC data}
\usage{
mw_wrangle_muac(
  df,
  sex,
  muac,
  age = NULL,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  .to = c("cm", "mm", "none"),
  .decimals = 3
)
}
\arguments{
\item{df}{A dataset object of class \code{data.frame} to wrangle data from.}

\item{sex}{A \code{numeric} or \code{character} vector of child's sex. Code values should
only be 1 or "m" for males and 2 or "f" for females. Make sure sex values
are coded in either of the aforementioned before to call the function. If input
codes are different than expected, the function will stop execution and
return an error message with the type of mismatch.}

\item{muac}{A vector of class \code{numeric} of child's age in months. If the class
is different than expected, the function will stop execution and return an error
message indicating the type of mismatch.}

\item{age}{A vector of class \code{numeric} of child's age in months.}

\item{.recode_sex}{Logical. Set to \code{TRUE} if the values for \code{sex} are not coded
as 1 (for males) or 2 (for females). Otherwise, set to \code{FALSE} (default).}

\item{.recode_muac}{Logical. Set to \code{TRUE} if the values for raw MUAC should be
converted to either centimeters or millimeters. Otherwise, set to \code{FALSE}
(default)}

\item{.to}{A choice of the unit to which the MUAC values should be converted;
"cm" for centimeters, "mm" for millimeters and "none" to leave as it is.}

\item{.decimals}{The number of decimals places the z-scores should have.
Default is 3.}
}
\value{
A data frame based on \code{df}. New variables named \code{mfaz} and
\code{flag_mfaz}, of child's MFAZ and detected outliers, will be created. When age
is not supplied, only \code{flag_muac} variable is created. This refers to outliers
detected based on the raw MUAC values.
}
\description{
Calculate z-scores for MUAC-for-age (MFAZ) and identify outliers based on
the SMART methodology. When age is not supplied, wrangling will consist only
in detecting outliers from raw the MUAC values.
}
\examples{
## When age is available ----
anthro.02 |>
  mw_wrangle_age(
    dos = NULL,
    dob = NULL,
    age = age,
    .decimals = 2
  ) |>
  mw_wrangle_muac(
    sex = sex,
    age = age,
    muac = muac,
    .recode_sex = TRUE,
    .recode_muac = TRUE,
    .to = "cm",
    .decimals = 3
  )

## When age is not available ----
anthro.02 |>
  mw_wrangle_muac(
    sex = sex,
    age = NULL,
    muac = muac,
    .recode_sex = TRUE,
    .recode_muac = TRUE,
    .to = "cm",
    .decimals = 3
  )

}
\references{
Bilukha, O., & Kianian, B. (2023). Considerations for assessment of measurement
quality of mid‐upper arm circumference data in anthropometric surveys and
mass nutritional screenings conducted in humanitarian and refugee settings.
\emph{Maternal & Child Nutrition}, 19, e13478. \url{https://doi.org/10.1111/mcn.13478}

SMART Initiative (2017). \emph{Standardized Monitoring and Assessment for Relief
and Transition}. Manual 2.0. Available at: \url{https://smartmethodology.org}.
}
\seealso{
\code{\link[=flag_outliers]{flag_outliers()}} \code{\link[=remove_flags]{remove_flags()}} \code{\link[=mw_wrangle_age]{mw_wrangle_age()}}
}
