% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_wfhz.R
\name{mw_wrangle_wfhz}
\alias{mw_wrangle_wfhz}
\title{Wrangle weight-for-height data}
\usage{
mw_wrangle_wfhz(df, sex, weight, height, .recode_sex = TRUE, .decimals = 3)
}
\arguments{
\item{df}{A dataset of class \code{data.frame} to wrangle data from.}

\item{sex}{A \code{numeric} or \code{character} vector of child's sex. Code values should
be 1 or "m" for boy and 2 or "f" for girl.}

\item{weight}{A vector of class \code{double} of child's weight in kilograms.}

\item{height}{A vector of class \code{double} of child's height in centimeters.}

\item{.recode_sex}{Logical. Set to \code{TRUE} if the values for \code{sex} are not coded
as 1 (for males) or 2 (for females). Otherwise, set to \code{FALSE} (default).}

\item{.decimals}{The number of decimals places the z-scores should have.
Default is 3.}
}
\value{
A data frame based on \code{df}. New variables named \code{wfhz} and
\code{flag_wfhz}, of child's weight-for-height z-scores and flags.
}
\description{
This function performs data wrangling by calculating the weight-for-height
followed by the detection and flagging of outliers based on the SMART flagging
criteria.
}
\details{
The flagging criterion used for the WFHZ is as in
\href{https://smartmethodology.org/}{SMART plausibility check}.
}
\examples{

anthro.01 |>
  mw_wrangle_wfhz(
    sex = sex,
    weight = weight,
    height = height,
    .recode_sex = TRUE,
    .decimals = 2
  )

}
