% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{flag_outliers}
\alias{flag_outliers}
\alias{remove_flags}
\title{Identify and flag outliers}
\usage{
flag_outliers(x, type = c("zscore", "crude"))

remove_flags(x, unit = c("zscore", "crude"))
}
\arguments{
\item{x}{A numeric vector holding either the WFHZ, the MFAZ values, or the
absolute MUAC values (in millimeters).}

\item{type}{The method you wish \code{flag_outliers()} to identify flag outliers
in the data. A choice between "zscore" (for WFHZ and MFAZ), and "crude" (for
absolute MUAC values).}

\item{unit}{A choice between "zscore" (for WFHZ and MFAZ), and "crude" (for
absolute MUAC values).}
}
\value{
A vector of the same length as input holding dummy values: 1 for is
a flag and 0 is not a flag.
}
\description{
Outliers are extreme values that deviate remarkably from the mean, making
them unlikely to be accurate measurements. \code{flag_outliers()} helps you to
identify them whether in the WFHZ, the MFAZ or the absolute MUAC values.
}
\details{
The flagging criteria for the WFHZ is as in
\href{https://smartmethodology.org/}{SMART plausibility check}. As for the MFAZ, it
uses the same criteria as WFHZ, whilst a fixed flagging criteria is used for
absolute MUAC values. This is as recommended by
\href{https://doi.org/10.1111/mcn.13478}{Bilukha, O., & Kianian, B. (2023).}
}
\examples{

## Sample data for absolute MUAC values ----
x <- c(90, 110, 140, 200, 119, 235)

## Apply `flag_outliers()` with type set to "crude" ----
flag_outliers(x, type = "crude")

## Sample data for MFAZ or for WFHZ values ----
x <- c(-2.265, -5.275, -0.72, -2.261, -2.264, -4.451, -2.261, -1.828)

# Apply `flag_outliers()` with type set to "zscore" ----
flag_outliers(x, type = "zscore")

}
