% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{flag_outliers}
\alias{flag_outliers}
\alias{remove_flags}
\title{Identify and flag outliers}
\usage{
flag_outliers(x, type = c("zscore", "crude"))

remove_flags(x, unit = c("zscore", "crude"))
}
\arguments{
\item{x}{A vector of class \code{double} of WFHZ or MFAZ or absolute MUAC values.
The latter should be in millimeters.}

\item{type}{A choice between \code{zscore} and \code{crude} for where outliers should be
detected and flagged from.}

\item{unit}{A choice between \code{zscore} and \code{crude} for where outliers should be
detected and flagged from.}
}
\value{
A vector of the same length as \code{x} of flagged observations that are
outliers: 1 for is a flag and 0 is not a flag.
}
\description{
Outliers are extreme values that deviate remarkably from the survey mean, making
them unlikely to be accurate measurements. This function detects and signals
them based on a criterion set for the WFHZ, the MFAZ and for the absolute MUAC
values.
}
\details{
The flagging criterion used for the WFHZ and the MFAZ is as in
\href{https://smartmethodology.org/}{SMART plausibility check}. A fixed flagging
criterion is used for the absolute MUAC values. This is as recommended by
\href{https://doi.org/10.1111/mcn.13478}{Bilukha, O., & Kianian, B. (2023).}
}
\examples{

## Sample data for absolute MUAC values ----
x <- anthro.01$muac

## Apply the function with type set to "crude" ----
flag_outliers(x, type = "crude")

## Sample data for MFAZ or for WFHZ values ----
x <- anthro.02$mfaz

# Apply the function with type set to "zscore" ----
flag_outliers(x, type = "zscore")

}
