% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_tables.R
\name{generate_pretty_table_mfaz}
\alias{generate_pretty_table_mfaz}
\alias{generate_pretty_table_wfhz}
\alias{generate_pretty_table_muac}
\title{Get a formatted and presentable output table for the plausibility checkers}
\usage{
generate_pretty_table_mfaz(df)

generate_pretty_table_wfhz(df)

generate_pretty_table_muac(df)
}
\arguments{
\item{df}{A summary table object of class \code{data.frame} returned by the
plausibility checkers.}
}
\value{
A \code{data.frame} as \code{df}. Columns are renamed, values formatted and
ready to be shared.
}
\description{
Useful to getting the output returned from the plausibility checkers
into a presentable format. It converts scientific notations to standard
notations, round values and rename columns to meaningful names.
}
\examples{

## Check the plausibility of WFHZ data ----

anthro.01 |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
) |>
check_plausibility_wfhz(
sex = sex,
age = age,
weight = weight,
height = height,
flags = flag_wfhz,
area = area
) |>
generate_pretty_table_wfhz()

## Check the plausibility of MUAC data ----

anthro.01 |>
process_muac_data(
sex = sex,
muac = muac,
age = NULL,
.recode_sex = TRUE,
.recode_muac = FALSE,
unit = "none"
) |>
check_plausibility_muac(
flags = flag_muac,
sex = sex,
muac = muac
) |>
generate_pretty_table_muac()

## Check the plausibility of MFAZ data ----

anthro.01 |>
mw_wrangle_age(
dos = dos,
dob = dob,
age = age
) |>
process_muac_data(
sex = sex,
age = "age",
muac = muac,
.recode_sex = TRUE,
.recode_muac = TRUE,
unit = "cm"
) |>
check_plausibility_mfaz(
flags = flag_mfaz,
sex = sex,
muac = muac,
age = age,
area = area
) |>
generate_pretty_table_mfaz()

}
