% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_muac.R
\name{mw_estimate_prevalence_muac}
\alias{mw_estimate_prevalence_muac}
\alias{mw_estimate_smart_age_wt}
\title{Estimate the prevalence of wasting based on MUAC for survey data}
\usage{
mw_estimate_prevalence_muac(df, wt = NULL, edema = NULL, .by = NULL)

mw_estimate_smart_age_wt(df, edema = NULL, .by = NULL)
}
\arguments{
\item{df}{A data set object of class \code{data.frame} to use. This must have been
wrangled using this package's wrangling function for MUAC data. Make sure
MUAC values are converted to millimeters after using the wrangler.
If this is not done, the function will stop execution and return an error message.
The function uses a variable name called \code{cluster} where the primary sampling unit IDs
are stored. Make sure the data set has this variable and its name has been
renamed to \code{cluster}, otherwise the function will error and terminate the execution.}

\item{wt}{A vector of class \code{double} of the final survey weights. Default is
\code{NULL} assuming a self weighted survey, as in the ENA for SMART software;
otherwise, when a vector of weights if supplied, weighted analysis is done.}

\item{edema}{A vector of class \code{character} of edema. Code should be
"y" for presence and "n" for absence of bilateral edema. Default is \code{NULL}.}

\item{.by}{A vector of class \code{character} or \code{numeric} of the geographical areas
or respective IDs for where the data was collected and for which the analysis
should be summarized at.}
}
\value{
A summarized table of class \code{data.frame} of the descriptive
statistics about wasting.
}
\description{
Calculate the prevalence estimates of wasting based on MUAC and/or bilateral
edema.
Before estimating, the function evaluates the quality of data by calculating
and rating the standard deviation of z-scores of muac-for-age (MFAZ) and the
p-value of the age ratio test; then it sets the analysis path that best fits
the data:
\itemize{
\item If all tests are rated as not problematic, a normal analysis is done.
\item If standard deviation is not problematic and age ratio test is problematic,
prevalence is age-weighted. This is to fix the likely overestimation of wasting
when there are excess of younger children in the data set.
\item If standard deviation is problematic and age ratio test is not, or both
are problematic,  analysis gets cancelled out and \code{NA}s get thrown.
}

Outliers are detected based on SMART flags on the MFAZ values and then
get excluded prior being piped into the actual prevalence analysis workflow.
}
\examples{
## When .by = NULL ----
mw_estimate_prevalence_muac(
  df = anthro.04,
  wt = NULL,
  edema = edema,
  .by = NULL
)

## When .by is not set to NULL ----
mw_estimate_prevalence_muac(
  df = anthro.04,
  wt = NULL,
  edema = edema,
  .by = province
)

## An application of `mw_estimate_smart_age_wt()` ----
.data <- anthro.04 |>
  subset(province == "Province 2")

mw_estimate_smart_age_wt(
  df = .data,
  edema = edema,
  .by = NULL
)

}
\references{
SMART Initiative (no date). \emph{Updated MUAC data collection tool}. Available at:
\url{https://smartmethodology.org/survey-planning-tools/updated-muac-tool/}
}
\seealso{
\code{\link[=mw_estimate_smart_age_wt]{mw_estimate_smart_age_wt()}} \code{\link[=mw_estimate_prevalence_mfaz]{mw_estimate_prevalence_mfaz()}}
\code{\link[=mw_estimate_prevalence_screening]{mw_estimate_prevalence_screening()}}
}
