% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prev_wasting_muac.R
\name{mw_estimate_prevalence_muac}
\alias{mw_estimate_prevalence_muac}
\alias{mw_estimate_smart_age_wt}
\title{Estimate the prevalence of wasting based on MUAC for survey data}
\usage{
mw_estimate_prevalence_muac(df, wt = NULL, edema = NULL, .by = NULL)

mw_estimate_smart_age_wt(df, edema = NULL, raw_muac = FALSE, .by = NULL)
}
\arguments{
\item{df}{A \code{tibble} object produced by \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}} and
\code{\link[=mw_wrangle_age]{mw_wrangle_age()}} functions. Note that MUAC values in \code{df}
must be in millimeters unit after using \code{\link[=mw_wrangle_muac]{mw_wrangle_muac()}}. Also, \code{df}
must have a variable called \code{cluster} which contains the primary sampling
unit identifiers.}

\item{wt}{A vector of class \code{double} of the survey sampling weights. Default
is NULL which assumes a self-weighted survey as is the case for a survey
sample selected proportional to population size (i.e., SMART survey sample).
Otherwise, a weighted analysis is implemented.}

\item{edema}{A \code{character} vector for presence of nutritional edema coded as
"y" for presence of nutritional edema and "n" for absence of nutritional
edema. Default is NULL.}

\item{.by}{A \code{character} or \code{numeric} vector of the geographical areas
or identifiers for where the data was collected and for which the analysis
should be summarised for.}
}
\value{
A summary \code{tibble} for the descriptive statistics about combined
wasting.
}
\description{
Estimate the prevalence of wasting based on MUAC and/or nutritional edema.
The function allows users to estimate prevalence in accordance with complex
sample design properties such as accounting for survey sample weights when
needed or applicable. The quality of the data is first evaluated by
calculating and rating the standard deviation of MFAZ and the p-value of the
age ratio test. Prevalence is calculated only when the standard deviation of
MFAZ is not problematic. If both standard deviation of MFAZ and p-value of
age ratio test is not problematic, straightforward prevalence estimation is
performed. If standard deviation of MFAZ is not problematic but p-value of
age ratio test is problematic, age-weighting is applied to prevalence
estimation to account for the over-representation of younger children in the
sample. If standard deviation of MFAZ is problematic, no estimation is done
and an NA value is returned. Outliers are detected based on SMART flagging
criteria for MFAZ. Identified outliers are then excluded before prevalence
estimation is performed.
}
\examples{
## When .by = NULL ----
mw_estimate_prevalence_muac(
  df = anthro.04,
  wt = NULL,
  edema = edema,
  .by = NULL
)

## When .by is not set to NULL ----
mw_estimate_prevalence_muac(
  df = anthro.04,
  wt = NULL,
  edema = edema,
  .by = province
)

## An application of `mw_estimate_smart_age_wt()` ----
.data <- anthro.04 |> subset(province == "Province 2")

mw_estimate_smart_age_wt(
  df = .data,
  edema = edema,
  .by = NULL
)

}
\references{
SMART Initiative (no date). \emph{Updated MUAC data collection tool}. Available at:
\url{https://smartmethodology.org/survey-planning-tools/updated-muac-tool/}
}
\seealso{
\code{\link[=mw_estimate_smart_age_wt]{mw_estimate_smart_age_wt()}} \code{\link[=mw_estimate_prevalence_mfaz]{mw_estimate_prevalence_mfaz()}}
\code{\link[=mw_estimate_prevalence_screening]{mw_estimate_prevalence_screening()}}
}
