% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_mfaz.R, R/prevalence_muac.R,
%   R/prevalence_wfhz.R
\name{compute_mfaz_prevalence}
\alias{compute_mfaz_prevalence}
\alias{compute_muac_prevalence}
\alias{mw_estimate_wfhz_prevalence}
\title{Compute the prevalence estimates of wasting on the basis of WFHZ, MFAZ or MUAC}
\usage{
compute_mfaz_prevalence(df, .wt = NULL, .edema = NULL, .summary_by = NULL)

compute_muac_prevalence(df, .wt = NULL, .edema = NULL, .summary_by = NULL)

mw_estimate_wfhz_prevalence(df, wt = NULL, edema = NULL, .by = NULL)
}
\arguments{
\item{df}{An already wrangled dataset object of class \code{data.frame} to use.}

\item{wt}{A vector of class \code{double} of the final survey weights. Default is
\code{NULL} assuming a self weighted survey, as in the ENA for SMART software;
otherwise, when a vector of weights if supplied, weighted analysis is computed.}

\item{edema}{A vector of class \code{character} of edema. Code should be
"y" for presence and "n" for absence of bilateral edema. Default is \code{NULL}.}

\item{.by}{A vector of class \code{character} of the geographical areas
where the data was collected and for which the analysis should be performed.}
}
\value{
A summarised table of class \code{data.frame} of the descriptive
statistics about wasting.
}
\description{
The prevalence is calculated in accordance with the complex sample design
properties inherent to surveys. This includes weighting the survey data where
applicable and applying PROBIT method estimation (for WFHZ) when the standard
deviation is problematic. This is as in the SMART Methodology.
}
\examples{

## An example of application of `compute_muac_prevalence()` ----

### When .summary.by = NULL ----

x <- compute_muac_prevalence(
df = anthro.04,
.wt = NULL,
.edema = edema,
.summary_by = NULL
)

print(x)

### When .summary_by is not set to NULL ----

p <- compute_muac_prevalence(
df = anthro.04,
.wt = NULL,
.edema = edema,
.summary_by = province
)

print(p)

## An example of application of `compute_wfhz_prevalence()` ----

### When .by = NULL ----
anthro.03 |>
  mw_wrangle_wfhz(
    sex = sex,
    weight = weight,
    height = height,
    .recode_sex = TRUE
  ) |>
  mw_estimate_wfhz_prevalence(
    wt = NULL,
    edema = edema,
    .by = NULL
  )

### When .by is not set to NULL ----

anthro.03 |>
  mw_wrangle_wfhz(
    sex = sex,
    weight = weight,
    height = height,
    .recode_sex = TRUE
  ) |>
  mw_estimate_wfhz_prevalence(
    wt = NULL,
    edema = edema,
    .by = district
  )

### When a weighted analysis is needed ----

anthro.02 |>
  mw_estimate_wfhz_prevalence(
    wt = "wtfactor",
    edema = edema,
    .by = province
  )

}
