% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_mfaz.R, R/prevalence_muac.R,
%   R/prevalence_wfhz.R
\name{compute_mfaz_prevalence}
\alias{compute_mfaz_prevalence}
\alias{compute_muac_prevalence}
\alias{compute_wfhz_prevalence}
\title{Compute the prevalence of wasting on the basis of WFHZ, MFAZ and MUAC}
\usage{
compute_mfaz_prevalence(df, .wt = NULL, .edema = NULL, .summary_by = NULL)

compute_muac_prevalence(df, .wt = NULL, .edema = NULL, .summary_by = NULL)

compute_wfhz_prevalence(df, .wt = NULL, .edema = NULL, .summary_by = NULL)
}
\arguments{
\item{df}{A data frame object with the required variables already wrangled.}

\item{.wt}{A numeric vector holding final survey weights. When set to \code{NULL},
the function assumes self weighted survey, as in the ENA for SMART software;
Otherwise when supplied, weighted analysis is computed.}

\item{.edema}{A character vector indicating if an observation has bilateral
edema or not. The codes are "y" for presence and "n" for absence of bilateral
edema. Default is \code{NULL}.}

\item{.summary_by}{A character vector containing data on the geographical areas
where the data was collected and for which the analysis should be performed at.}
}
\value{
A table with the descriptive statistics about wasting.
}
\description{
The prevalence is calculated in accordance with the complex sample design
properties inherent to surveys. This includes weighting the survey data where
applicable and applying PROBIT method estimation of prevalence (for WFHZ and
MFAZ) when standard deviation is problematic. This is as in the SMART Methodology.
}
\examples{

## An example of application of `compute_muac_prevalence()` ----

### When .summary.by = NULL ----

x <- compute_muac_prevalence(
df = anthro.04,
.wt = NULL,
.edema = edema,
.summary_by = NULL
)

print(x)

### When .summary_by is not set to NULL ----

p <- compute_muac_prevalence(
df = anthro.04,
.wt = NULL,
.edema = edema,
.summary_by = province
)

print(p)

## An example of application of `compute_wfhz_prevalence()` ----

### When .summary_by = NULL ----
anthro.03 |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
) |>
compute_wfhz_prevalence(
.wt = NULL,
.edema = edema,
.summary_by = NULL
)

### When .summary_by is not set to NULL ----

anthro.03 |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
) |>
compute_wfhz_prevalence(
.wt = NULL,
.edema = edema,
.summary_by = district
)

### When a weighted analysis is needed ----

anthro.02 |>
compute_wfhz_prevalence(
.wt = "wtfactor",
.edema = edema,
.summary_by = province
)

}
