% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevalence_wfhz.R
\name{apply_probit_approach}
\alias{apply_probit_approach}
\alias{compute_probit_prevalence}
\title{Compute the prevalence of wasting on the basis of the PROBIT method.}
\usage{
apply_probit_approach(x, .status = c("gam", "sam"))

compute_probit_prevalence(df, .summary_by = NULL, .for = c("wfhz", "mfaz"))
}
\arguments{
\item{x}{A numeric vector holding WFHZ or MFAZ values.}

\item{.status}{A choice on the form of wasting for which the prevalence should
be calculated for.}

\item{df}{A data frame object with the required variables already wrangled.}

\item{.summary_by}{A character vector containing data on the geographical areas where
the data was collected. This is to group the survey design object into different
geographical areas in the data and allow for summaries to be computed for each of them.
Default is NULL.}

\item{.for}{A choice between "wfhz" and "mfaz" for the anthropometric index you wish
to compute PROBIT prevalence on.}
}
\value{
A data frame with the prevalence. No confidence intervals are
yielded.
}
\description{
This approach is applied when the standard deviation of WFHZ is problematic.
The PROBIT method estimates the prevalence of wasting indirectly by calculating
the area under the curve the tail of the curve, from negative infinitive to
the given threshold, using the cumulative normal distribution function with
the mean and standard deviation as inputs.
}
