% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age.R
\name{process_age}
\alias{process_age}
\title{Process child's age}
\usage{
process_age(df, svdate = NULL, birdate = NULL, age)
}
\arguments{
\item{df}{A dataset of class \code{data.frame} to process age from.}

\item{svdate}{A vector of class \code{Date} for date of data collection.
Default is \code{NULL}.}

\item{birdate}{A vector of class \code{Date} for child's date of birth.
Default is \code{NULL}.}

\item{age}{A vector of class \code{integer} of age in months, usually estimated
using local event calendars.}
}
\value{
A \code{data.frame} based on \code{df}. The variable \code{age} that is required to be
included in \code{df} will be filled where applicable with the age in months for
each row of data in \code{df}. A new variable for \code{df} named \code{age_days} will be
created. Values for \code{age} and \code{age_days} for children less than 6.0 and greater
than or equal to 60.0 months old will be set to \code{NA}.
}
\description{
Process child's age for downstream analysis. This includes calculating age
in months based on the date of data collection and child's date of birth and
setting to \code{NA} the age values that are less than 6.0 and greater than or equal
to 60.0 months old.
}
\examples{

## A sample data ----
df <- data.frame(
survy_date = as.Date(c(
"2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01")),
birthdate = as.Date(c(
"2019-01-01", NA, "2018-03-20", "2019-11-05", "2021-04-25")),
age = c(NA, 36, NA, NA, NA)
)

## Apply the function ----
df |>
process_age(
svdate = "survy_date",
birdate = "birthdate",
age = age
)

}
