% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_raters.R
\name{rate_overall_quality}
\alias{rate_overall_quality}
\title{Rate the overall acceptability of the data}
\usage{
rate_overall_quality(q)
}
\arguments{
\item{q}{A \code{numeric} or \code{integer} vector of data acceptability scores.}
}
\value{
A vector of class \code{factor} with the same length as \code{q} of overall
rate of acceptability of the data.
}
\description{
Rate the overall data acceptability score into "Excellent", "Good",
"Acceptable" or "Problematic".
}
\keyword{internal}
