% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_raters.R
\name{rate_overall_quality}
\alias{rate_overall_quality}
\title{Rate the overall data acceptability score}
\usage{
rate_overall_quality(q)
}
\arguments{
\item{q}{A vector of class \code{numeric} or \code{integer} of data acceptability scores.
If the class does not match the expected type, the function
will stop execution and return an error message indicating the type of mismatch.}
}
\value{
A vector of class \code{factor} of the same length as \code{q}, providing an overall
rating of data acceptability.
}
\description{
Rate the overall data acceptability score into "Excellent", "Good", "Acceptable"
or "Problematic".
}
\keyword{internal}
