% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_raters.R
\name{rate_std}
\alias{rate_std}
\title{Rate the acceptability of the standard deviation}
\usage{
rate_std(sd, .of = c("zscores", "raw_muac"))
}
\arguments{
\item{sd}{A vector of class \code{double}, containing values of the standard deviation
from the data set. If the class does not match the expected type, the function
will stop execution and return an error message indicating the type of mismatch.}

\item{.of}{Specifies the data set where the rating should be done, with options:
"wfhz", "mfaz", or "raw_muac".}
}
\value{
A vector of class \code{factor} of the same length as input, for the
acceptability rate.
}
\description{
Rate the acceptability of the standard deviation of WFHZ, MFAZ, and raw MUAC data.
Rating follows the SMART methodology criteria.
}
\keyword{internal}
