% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_muac}
\alias{recode_muac}
\title{Convert MUAC values to either centimeters or millimeters}
\usage{
recode_muac(x, .to = c("cm", "mm"))
}
\arguments{
\item{x}{A vector of the raw MUAC values. The class can either be
\code{double} or \code{numeric}.}

\item{.to}{A choice of the unit to which the MUAC values should be converted.}
}
\value{
A numeric vector of the same length as \code{x}, with values converted
to the chosen unit.
}
\description{
Convert MUAC values to either centimeters or millimeters as required.
}
\examples{

## Recode from millimeters to centimeters ----
muac <- anthro.01$muac
muac_cm <- recode_muac(muac, .to = "cm")

## Using the `muac_cm` object to recode it back to "mm" ----
muac_mm <- recode_muac(muac_cm, .to = "mm")

}
