% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_muac}
\alias{recode_muac}
\title{Convert MUAC values to either centimeters or millimeters}
\usage{
recode_muac(x, .to = c("cm", "mm"))
}
\arguments{
\item{x}{A vector of the raw MUAC values. The class can either be
\code{double} or \code{numeric} or \code{integer}. If different than expected, the function
will stop execution and return an error message indicating the type of mismatch.}

\item{.to}{The unit to convert MUAC values to. Defaults to \code{cm}. If set to \code{cm},
the function checks if the values are in \code{mm}; if not, it throws an error.
Similarly, if set to \code{mm}, the function ensures the values are in \code{cm} before
to proceed.}
}
\value{
A numeric vector of the same length as \code{x}, with values converted
to the chosen unit.
}
\description{
Convert MUAC values to either centimeters or millimeters as required.
Before to covert, the function checks if the supplied MUAC
values are in the opposite unit of the intended conversion. If not,
execution stops and an error message is returned.
}
\examples{
## Recode from millimeters to centimeters ----
muac <- anthro.01$muac
muac_cm <- recode_muac(muac, .to = "cm")

## Using the `muac_cm` object to recode it back to "mm" ----
muac_mm <- recode_muac(muac_cm, .to = "mm")

}
