% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_muac}
\alias{recode_muac}
\title{Convert MUAC values to either centimeters or millimeters}
\usage{
recode_muac(x, .to = c("cm", "mm"))
}
\arguments{
\item{x}{A vector of the raw MUAC values. The class can either be
\code{double} or \code{numeric} or \code{integer}. If different than expected, the function
will stop execution and return an error message indicating the type of mismatch.}

\item{.to}{A choice between \code{cm} (centimeters) and \code{mm} (millimeters) for the
measuring unit to convert MUAC values to. Before to execute the conversion,
the function checks if values are in the opposite unit; in case not, the
execution stops and an error message is returned. Strive to address the error
and try again.}
}
\value{
A \code{numeric} vector of the same length as \code{x}, with values converted
to the chosen measuring unit.
}
\description{
Convert MUAC values to either centimeters or millimeters as required.
Before to covert, the function checks if the supplied MUAC
values are in the opposite unit of the intended conversion. If not,
execution stops and an error message is returned.
}
\examples{
## Recode from millimeters to centimeters ----
muac <- anthro.01$muac
muac_cm <- recode_muac(muac, .to = "cm")

## Using the `muac_cm` object to recode it back to "mm" ----
muac_mm <- recode_muac(muac_cm, .to = "mm")

}
