% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{recode_muac}
\alias{recode_muac}
\title{Convert MUAC values to either centimeters or millimeters}
\usage{
recode_muac(muac, unit = c("cm", "mm"))
}
\arguments{
\item{muac}{A numeric vector holding the absolute MUAC values.}

\item{unit}{A choice of the unit to which you wish to convert the MUAC
values into.}
}
\value{
A numeric vector of the same length as input, with values converted
into your chosen unit.
}
\description{
Recode the MUAC values into either centimeters or millimeters as required.
\code{recode_muac()} works inside \code{\link[dplyr:mutate]{dplyr::mutate()}} or \code{\link[base:transform]{base::transform()}}.
}
\examples{

## A sample of MUAC data in millimeters ----
muac <- seq(90, 250, by = 4)

## Apply the function ----
recode_muac(muac, unit = "cm")

## A sample of MUAC data in centimeters ----
muac <- seq(9.0, 25.0, by = 0.2)

# Apply the function ----
recode_muac(muac, unit = "mm")

}
