% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_muac}
\alias{recode_muac}
\title{Convert MUAC values to either centimeters or millimeters}
\usage{
recode_muac(x, .to = c("cm", "mm"))
}
\arguments{
\item{x}{A vector of raw MUAC values. The class can either be \code{double} or
\code{numeric} or \code{integer}.}

\item{.to}{Either "cm" (centimeters) or "mm" (millimeters) for the unit of
measurement to convert MUAC values to.}
}
\value{
A \code{numeric} vector of the same length as \code{x} with values set to
specified unit of measurement.
}
\description{
Convert MUAC values to either centimeters or millimeters
}
\examples{
## Recode from millimeters to centimeters ----
muac_cm <- recode_muac(
  x = anthro.01$muac,
  .to = "cm"
)
head(muac_cm)

## Using the `muac_cm` object to recode it back to "mm" ----
muac_mm <- recode_muac(
  x = muac_cm,
  .to = "mm"
)
tail(muac_mm)

}
