% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scorers.R
\name{score_overall_quality}
\alias{score_overall_quality}
\title{Get the overall acceptability score from the acceptability rate scores}
\usage{
score_overall_quality(
  cl_flags,
  cl_sex,
  cl_age,
  cl_dps_m = NULL,
  cl_dps_w = NULL,
  cl_dps_h = NULL,
  cl_std,
  cl_skw,
  cl_kurt,
  .for = c("wfhz", "mfaz")
)
}
\arguments{
\item{.for}{A choice between "wfhz" and "mfaz" for the type of scorer to
apply. Default is "wfhz".}
}
\value{
A \code{numeric} value for the overall data quality (acceptability)
score.
}
\description{
Get the overall acceptability score from the acceptability rate scores
}
\keyword{internal}
