% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scorers.R
\name{assign_penalty_points_flags_and_sd}
\alias{assign_penalty_points_flags_and_sd}
\alias{assign_penalty_points_age_sex_ratio}
\alias{assign_penalty_points_skew_kurt}
\title{Score the rating of proportion of flagged data, the magnitude of the standard
deviation, skewness, kurtosis and the p-values sex and age ratio test}
\usage{
assign_penalty_points_flags_and_sd(x)

assign_penalty_points_age_sex_ratio(x)

assign_penalty_points_skew_kurt(x)
}
\arguments{
\item{x}{A character vector holding the test classifications for the proportion
of flagged data, the magnitude of the standard deviation, the p-values of the
age and sex ratio tests, as well as the results of skewness and kurtosis tests.}
}
\value{
A numeric vector with the corresponding score.
}
\description{
\code{assign_penalty_points_flags_and_sd()} ranks the proportion of the flagged
values in the data and the magnitude of standard deviation based on the SMART
scoring criteria.
}
\details{
The ranking is as in \href{https://smartmethodology.org/}{SMART Plausibility checks}.
}
