% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_scorers.R
\name{assign_penalty_points_flags_and_sd}
\alias{assign_penalty_points_flags_and_sd}
\alias{assign_penalty_points_age_sex_ratio}
\alias{assign_penalty_points_skew_kurt}
\title{Score the acceptability classification of the standard deviation and percentage
of flagged data test results}
\usage{
assign_penalty_points_flags_and_sd(x)

assign_penalty_points_age_sex_ratio(x)

assign_penalty_points_skew_kurt(x)
}
\arguments{
\item{x}{A vector of class \code{character} of acceptability classification of the
plausibility test results.}
}
\value{
A vector of class \code{integer} of the same length as \code{x} for the score.
}
\description{
Attribute a penalty point based on the acceptability classification in which
the plausibility test result falls.
}
\details{
The scoring criteria is as in \href{https://smartmethodology.org/}{SMART Plausibility checks}.
}
