% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wranglers.R
\name{process_muac_data}
\alias{process_muac_data}
\alias{process_wfhz_data}
\title{Process and censor weight-for-height and MUAC data}
\usage{
process_muac_data(
  df,
  sex,
  muac,
  age = NULL,
  .recode_sex = TRUE,
  .recode_muac = TRUE,
  unit = c("cm", "mm", "none")
)

process_wfhz_data(df, sex, weight, height, .recode_sex = TRUE)
}
\arguments{
\item{df}{The input data frame with the required variables.}

\item{sex}{A numeric or character vector of child's sex. Code values should
either be 1 or "m" for boy and 2 or "f" for girl. The variable name must be
sex, otherwise it will not work.}

\item{muac}{A numeric vector holding the absolute MUAC values.}

\item{age}{A numeric vector of child's age in months. It must be named age,
otherwise it will not work. For instance, if given as following: age = months
it will not work.}

\item{.recode_sex}{Logical. It asks whether sex should be recoded. In the end,
the variable sex have values coded as 1 for boy and 2 for girl. Setting
\code{.recode_sex = TRUE} works over "m" and "f" values. If your vector is coded
differently, make sure to put it in "m" and "f" or in 1 or 2 right away.}

\item{.recode_muac}{Logical. Choose between \code{TRUE} if you wish to recode
the MUAC values into either centimeters or millimeters.}

\item{unit}{A choice of the unit to which you wish to convert the MUAC
variable into. Choose "cm" for centimeters, "mm" for millimeters and "none"
to leave as it is.}

\item{weight}{A numeric vector holding the weight values of the child in
kilograms.}

\item{height}{A numeric vector holding the height values of the child in
centimeters.}
}
\value{
A data frame of the same length as the input with additional
columns: one named \code{wfhz} or \code{mfaz} that holds the zscore values, and the other
holding dummy values: 1 (is a flag) and 0 (is not a flag). For the
\code{process_muac_data} function, when age is not supplied, only \code{flag_muac} is
added. This refers to flags based based on absolute MUAC values as recommended by
\href{https://doi.org/10.1111/mcn.13478}{Bilukha, O., & Kianian, B. (2023).}.
}
\description{
This is the job of \code{process_wfhz_data} and \code{process_muac_data()}. They are
responsible for computing the weight-for-height and the muac-for-age z-scores
respectively, and censor the data by flagging outliers based on the SMART flags.
For the latter, if age is not supplied, the function censors the absolute MUAC
values.
}
\examples{

## An example application of `process_wfhz_data()` ----

anthro.01 |>
process_wfhz_data(
sex = sex,
weight = weight,
height = height,
.recode_sex = TRUE
)

## An example application of `process_muac_data()` ----

### Sample data ----
df <- data.frame(
 survey_date = as.Date(c(
 "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01", "2023-01-01")),
 birthdate = as.Date(c(
 "2019-01-01", NA, "2018-03-20", "2019-11-05", "2021-04-25")),
 age = c(NA, 36, NA, NA, NA),
 sex = c("m", "f", "m", "m", "f"),
 muac = c(110, 130, 300, 123, 125)
 )

 ### The application of the function ----
 df |>
 process_age(
 svdate = "survey_date",
 birdate = "birthdate",
 age = age
 ) |>
 process_muac_data(
 sex = sex,
 age = "age",
 muac = muac,
 .recode_sex = TRUE,
 .recode_muac = TRUE,
 unit = "cm"
 )

}
